Control.Foldl.End

Description

This module provides a version of Foldl that can return early

Data type declarations

newtype FoldE a b

Fold left or return an early result.

Normal Folds are strict and thus loop on infinite input:

 λ> F.fold F.head ([1..] :: [Int])
 ^CInterrupted.

 λ> foldE headE ([1..] :: [Int])
 Just 1
 (0.05 secs, 450,312 bytes)

Since FoldE returns early upon success, it can terminate on infinite inputs.

Can be composed in at least three ways!

See the Semigroupoid instance for the first way: vertical composition.

Constructors

FoldE

Fields

runFoldE :: FoldM (Either b) a b

Data type helpers

withFoldE :: (FoldM (Either b) a b -> FoldM (Either d) c d) -> FoldE a b -> FoldE c d

Unwrap and wrap the FoldE newtype

fmapAlternative :: (b -> c) -> FoldE a b -> FoldE a c

TODO: compare performance to fmap

stepFoldE :: a -> FoldE a b -> FoldE a b

stepFoldM, specialized to FoldE

unwrapFoldE :: FoldE a b -> Either b (FoldE a b)

unwrapFoldM specialized to FoldE

Folds

takeE :: Int -> FoldE a [a]

Take n and return a list

 λ> foldE (toFoldE F.list o takeE 3) [1..10]
 [[1,2,3],[4,5,6],[7,8,9],[10]]

dropE :: Int -> FoldE a ()

Drop n and return nothing

 λ> foldE (toFoldE F.list o (dropE 3 >> headE)) [1..10]
 [Just 5,Just 9,Nothing]

lineE :: FoldE Char String

Take a single line (and toss its end, if it has one)

headE :: FoldE a (Maybe a)

Return the first value and always fail if extracted earlier.

I'm not really sure why the values are being duplicated.. but it's probably because the first early return is assumed to have _not_ consumed the last value it reached..

 λ> foldE (toFoldE F.list o headE) [1..3]
 [Just 1,Just 1,Just 2,Just 2,Just 3,Just 3,Nothing]

tailE :: FoldE a ()

Drop the first element provided or return `()` if extracted.

lastE :: FoldE a (Maybe a)

Return Just the last element or Nothing

product :: (Num a, Eq a) => FoldE a a

Returns early on `0`.

Use toFoldE with the product from Control.Foldl to multiply all elements.

Note: will likely be less performant if you're unlikely to encounter 0.

elem :: Eq a => a -> FoldE a Bool

Is the given value an element?

This is preetty slow.. upwards of 30x slower than elem

elem' :: Eq a => a -> FoldE a Bool

Wow, this is still pretty slow. (It's about 70% faster with lists of around 2^20 elements). Maybe if it's inlined and -O3'd?

find :: (a -> Bool) -> FoldE a (Maybe a)

Either return Just the first value matching the predicate or Nothing.

index :: Int -> FoldE a (Maybe a)

Either return Just the element at the given index (`0` at the beginning of the FoldE) or Nothing.

lookup :: Eq a => a -> FoldE (a, b) (Maybe b)

Return Just the value associated with the given key or Nothing

minimum :: (Bounded a, Ord a) => FoldE a (Maybe a)

Returns early on minBound.

Use toFoldE with the minimum for the general case.

Likely to be much less performant if you don't expect to have many minBounds in what you're folding.

maximum :: (Bounded a, Ord a) => FoldE a (Maybe a)

Returns early on maxBound.

Likely to be much less performant if you don't expect to have many maxBounds in what you're folding.

elemIndex :: Eq a => a -> FoldE a (Maybe Int)

Just the first index of the given value or Nothing

findIndex :: (a -> Bool) -> FoldE a (Maybe Int)

Find the first index for which the predicate returns True

genericIndex :: Integral i => i -> FoldE a (Maybe a)

index for any Integral

Ordered folds

lookupOrd :: Ord a => a -> FoldE (a, b) (Maybe b)

Assumes the input keys are monotonically increasing (inc)

elemIndexOrd :: Ord a => a -> FoldE a (Maybe Int)

Assumes the inputs are monotonically increasing (inc)

Conversion

fromFoldE :: FoldE a b -> Fold a b

Convert from a FoldE to a Fold by binding (>>=) steps together and using joinEither on the result.

toFoldE :: Fold a b -> FoldE a b

Never returns Left. `fromFoldE . toFoldE == id`

foldE1 :: FoldE a b -> a -> b

foldE1 specialized to Identity and pre-applied for convenience.

toCofreeReader :: FoldE a b -> Cofree (ReaderT a (Either b)) b

Unfold the successive applications of a FoldE into a Cofree ReaderT

foldEToFree :: FoldE a b -> FreeFoldM (Either b) a b

Convert a FoldE to a FreeFoldM using toCofreeReader

Folding

foldE :: Foldable t => FoldE a b -> t a -> b

Run a FoldE over a Foldable

Combinators

foldEThenE :: FoldE a b -> (b -> FoldE a c) -> FoldE a c

Fold with the first fold until it returns Left, then feed its input into the second fold and continue.

foldEThenFree :: FoldE a b -> (b -> FoldE a c) -> FreeFoldM (Either c) a c

Fold with the first FoldE until Left, then use its result to construct the second fold and continue.

foldEThenERec :: FoldE a b -> (b -> FoldE a c) -> FoldE a c

So, I was afraid that this wouldn't be more performant, but here it is:

λ> (foldE $ foldEThenE (takeE 3) (x -> (length x +) $ toFoldE F.sum)) $ [1..2^20]
(2.77 secs, 1,771,691,168 bytes)

λ> (foldE $ foldEThenE (takeE 3) (x -> (length x +) $ toFoldE F.sum)) $ [1..2^21]
(5.41 secs, 3,542,965,480 bytes)

λ> (foldE $ foldEThenERec (takeE 3) (x -> (length x +) $ toFoldE F.sum)) $ [1..2^20]
(0.68 secs, 857,335,424 bytes)

λ> (foldE $ foldEThenERec (takeE 3) (x -> (length x +) $ toFoldE F.sum)) $ [1..2^21]
(1.50 secs, 1,714,251,384 bytes)

Of course we'll have to wait for criterion benchmarks to really tell.

foldEThenERec' :: FoldE a b -> (b -> FoldE a c) -> FoldE a c

foldEThenERec, using foldFreeThenFree instead of directly recursing.

Preliminary tests show it may be a hair faster than foldEThenERec, but we'll have to wait for criterion tests to really tell.

foldEThenE_ :: FoldE a b -> FoldE a c -> FoldE a c

foldEThenE, but drop the first result

foldEThen :: FoldE a b -> (b -> Fold a c) -> Fold a c

foldEThenE, but never fail with the second fold

foldEThen_ :: FoldE a b -> Fold a c -> Fold a c

foldEThenE_, but never fail on the second fold

peekLast :: FoldE a b -> (a -> b -> c) -> FoldE a (Either b c)

Ok, I think this is probably _not_ very efficient, but here goes:

Example:

 peekLast `(,)` :: FoldE a b -> FoldE a (Either b (a, b))
 peekLast `(,)` head :: FoldE a (Either (Maybe a) (a, Maybe a))

 λ> peekLast headE (,) foldE [1..1]
 Right (1,Just 1)
 (0.02 secs, 549,328 bytes)
 λ> peekLast headE (,) foldE [1..10]
 Right (1,Just 1)
 (0.01 secs, 549,328 bytes)

peekLast_ :: FoldE a b -> (a -> c) -> FoldE a (Either b c)

peekLast, discarding the result of the FoldE

foldE2 :: (a -> b) -> (a -> a -> b) -> FoldE b c -> FoldE a c

Given a fold over the results of cases for one or and maybe the previous element, produce a fold over arbitrary inputs.

  diffs :: Num a => FoldE a (Maybe a)
  diffs = foldE2 ifOne ifMultiple _
   where
     ifOne = Just
     ifMultiple = liftM2 (flip (-))

I've found it quite helpful for folding over pairs of elements.

foldERec2 :: (a -> b) -> (a -> a -> b) -> FoldE b c -> FoldE a c

foldE2 using foldFree2

Predicate combinators

countPrefixE :: Enum i => (a -> Bool) -> FoldE a i

Count the number of elements in the prefix that satisfy the predicate. Should be equivalent to: foldWhile p length

takeWhileE :: (a -> Bool) -> FoldE a [a]

- We can use this with other combinators to build folding parsers.

Warning: composing two equivalent takeWhileE's will diverge

foldWhile :: (a -> Bool) -> Fold a b -> FoldE a b

Fold until the predicate fails

foldUntil :: (a -> Bool) -> Fold a b -> FoldE a b

Fold until the predicate succeeds

foldWhileE :: (a -> Bool) -> FoldE a b -> FoldE a b

Fold until the given fold returns early or the predicate fails

foldUntilE :: (a -> Bool) -> FoldE a b -> FoldE a b

Fold until the given fold returns early or the predicate succeeds

foldWhile2 :: (a -> Bool) -> (a -> a -> Bool) -> Fold a b -> FoldE a b

If there's no previous element, equivalent to foldWhile on the first predicate. If there's a previous element, then apply to it and the current element:

 predicate2 previous current

bifoldWhile :: (a -> Bool) -> (a -> Bool) -> Fold a b -> FoldE a b

If there's no previous element, equivalent to foldWhile on the first predicate. If there's a previous element, then we do:

 predicate1 previous && predicate2 current

I believe it can be more efficient than foldWhile2, but of course your predicate has to be separable.

foldWhileE2 :: (a -> Bool) -> (a -> a -> Bool) -> FoldE a b -> FoldE a b

foldWhile2, but the FoldE can return early

bifoldWhileE :: (a -> Bool) -> (a -> Bool) -> FoldE a b -> FoldE a b

bifoldWhile, but the FoldE can return early

skipWhile :: (a -> Bool) -> FoldE a ()

Skip while the predicate holds.

I'm not sure whether this should have one of the types:

 FoldE a a
 FoldE a (Maybe a)

To allow it to return the element that the predicate fails on.

Should be equivalent to:

 void . foldWhile (pure ())

skipUntil :: (a -> Bool) -> FoldE a ()

Skip until the predicate holds

satisfyE :: (a -> Bool) -> FoldE a (Maybe a)

If the predicate passes, return the value. If folded over nothing, return Nothing.

prefilterMaybe :: (a -> Maybe b) -> FoldE b c -> FoldE a c

Apply the function, ignoring the Nothings

Free predicate combinators

freeToFoldE :: FreeFoldM (Either b) a b -> FoldE a b

Convert a FreeFoldM to a FoldE using fromCofreeReader

bifoldWhileE' :: (a -> Bool) -> (a -> Bool) -> FoldE a b -> FoldE a b

bifoldWhileE implemented in terms of Cofree (ReaderT a (Either b)) b.

bifoldWhileFreeE :: (a -> Bool) -> (a -> Bool) -> Cofree (ReaderT a (Either b)) b -> Cofree (ReaderT a (Either b)) b

Since we can map over a single layer of a Cofree, returning Left if the first predicate fails, we can apply the second predicate using foldWhileFreeE.

This avoids the need for Maybe in the state variable.

foldWhileFreeE :: (a -> Bool) -> Cofree (ReaderT a (Either b)) b -> Cofree (ReaderT a (Either b)) b

See foldWhileE

Container predicates

null :: FoldE a Bool

Are there any elements?

and :: FoldE Bool Bool

Returns True if all elements are True, False otherwise

or :: FoldE Bool Bool

Returns True if any elements are True, False otherwise

all :: (a -> Bool) -> FoldE a Bool

Returns True if all elements satisfy the predicate, False otherwise

any :: (a -> Bool) -> FoldE a Bool

Returns True if any elements satisfy the predicate, False otherwise

Series predicates

eq :: Eq a => FoldE a Bool

Are all elements equal?

inc :: Ord a => FoldE a Bool

Monotonically increasing

inc' :: Ord a => FoldE a Bool

Strictly increasing

dec :: Ord a => FoldE a Bool

Monotonically decreasing

dec' :: Ord a => FoldE a Bool

Strictly decreasing