Control.Foldl.Free

Description

This module provides a version of Foldl in terms of Cofree and ReaderT that's isomorphic to Foldl.

Data types

newtype FreeFoldM m a b

By expressing the state of a left fold as a Cofree:

 Cofree (current state, new value -> monadic context (new state, new new value -> ..))

we can express arbitrary recursions, one step at a time, and allowing things like:

 If even x then replace all further functions with const 0.

This allows us ~(in theory)~ to much more efficiently express combinators such as foldEThenE.

Note: this type is isomorphic to FoldM.

Constructors

FreeFoldM

Fields

runFreeFoldM :: m (Cofree (ReaderT a m) b)

type FreeFold = FreeFoldM Identity

A pure FreeFoldM

Construction and destruction helpers

withFreeFoldM :: (m (Cofree (ReaderT a m) b) -> n (Cofree (ReaderT c n) d)) -> FreeFoldM m a b -> FreeFoldM n c d

Helper to abstract-out wrapping and unwrapping a FreeFoldM

withFreeFoldM2 :: (m1 (Cofree (ReaderT a1 m1) b1) -> m2 (Cofree (ReaderT a2 m2) b2) -> m3 (Cofree (ReaderT a3 m3) b3)) -> FreeFoldM m1 a1 b1 -> FreeFoldM m2 a2 b2 -> FreeFoldM m3 a3 b3

Helper to abstract-out wrapping and unwrapping a FreeFoldM

asFreeFoldM2 :: (FreeFoldM m1 a1 b1 -> FreeFoldM m2 a2 b2 -> FreeFoldM m3 a3 b3) -> m1 (Cofree (ReaderT a1 m1) b1) -> m2 (Cofree (ReaderT a2 m2) b2) -> m3 (Cofree (ReaderT a3 m3) b3)

Helper to abstract-out wrapping and unwrapping a pair of FreeFoldM's

Expose the underlying Cofree

extractFreeFoldM :: Functor m => FreeFoldM m a b -> m b

Extract the first result from a FreeFoldM

unwrapFreeFold :: FreeFold a b -> Reader a (FreeFold a b)

Expose a layer of a FreeFold

unwrapFreeFoldM :: Monad m => FreeFoldM m a b -> ReaderT a m (FreeFoldM m a b)

Expose a layer of a FreeFoldM

wrapFreeFoldM :: Monad m => m (FreeFoldM m a b) -> FreeFoldM m a b

Wrap a Monadic layer around a FreeFoldM

Conversion

freeFold :: FreeFold a b -> Fold a b

Construct a Fold from a FreeFold.

(This is the the inverse of toFreeFold.)

freeFoldM :: Monad m => FreeFoldM m a b -> FoldM m a b

Construct a FoldM from a FreeFoldM.

(This is the the inverse of toFreeFoldM.)

toFreeFold :: Fold a b -> FreeFold a b

Convert a Fold to a FreeFold.

(This is the inverse of freeFold.)

toFreeFoldM :: Monad m => FoldM m a b -> FreeFoldM m a b

Convert a FoldM to a FreeFoldM.

(This is the inverse of freeFoldM.)

Modification

stepFreeFold :: a -> FreeFold a b -> FreeFold a b

Perform a single step of a FreeFold

stepFreeFoldM :: Monad m => a -> FreeFoldM m a b -> FreeFoldM m a b

Perform a single step of a FreeFoldM

skipFreeFold :: FreeFold a b -> FreeFold a b

Skip the first step of a FreeFold

skipFreeFoldM :: Monad m => FreeFoldM m a b -> FreeFoldM m a b

Skip the first step of a FreeFoldM

Combinators

foldFreeThenFree :: FreeFoldM (Either b) a b -> (b -> FreeFoldM (Either c) a c) -> FreeFoldM (Either c) a c

Fold with the first until Left, then use its result to construct the second fold and continue.

foldCofreeThenCofree :: Cofree (ReaderT a (Either b)) b -> (b -> Either c (Cofree (ReaderT a (Either c)) c)) -> Either c (Cofree (ReaderT a (Either c)) c)

Fold with the first, extracting each result to provide to the Cofree, until Left, then use its result to construct the second Cofree ReaderT and continue.

foldFreeThenFreeM :: Monad m => FreeFoldM (ExceptT b m) a b -> (b -> FreeFoldM (ExceptT c m) a c) -> FreeFoldM (ExceptT c m) a c

foldFreeThenFree generalized to ExceptT

foldCofreeThenCofreeM :: Monad m => Cofree (ReaderT a (ExceptT b m)) b -> (b -> ExceptT c m (Cofree (ReaderT a (ExceptT c m)) c)) -> ExceptT c m (Cofree (ReaderT a (ExceptT c m)) c)

foldCofreeThenCofree generalized to ExceptT

foldFree2 :: (a -> b) -> (a -> a -> b) -> FreeFoldM (Either c) b c -> FreeFoldM (Either c) a c

Given cases for one or (previous, current) elements and a fold over the results, produce a combined FreeFoldM.

foldCofree2 :: (a -> b) -> (a -> a -> b) -> Cofree (ReaderT b (Either c)) c -> Cofree (ReaderT a (Either c)) c

Given cases for one or (previous, current) elements and a fold over the results, produce a combined fold.

unpairCofreeReader :: Monad m => Cofree (ReaderT (a, a) m) b -> ReaderT a m (Cofree (ReaderT a m) b)

Resolve paired to unpaired inputs by prepending a ReaderT

unpairFoldFree :: Monad m => a -> FreeFoldM m (a, a) b -> FreeFoldM m a b

Given an initial value, we can convert a FreeFoldM of sucessive (previous, current) pairs to a simple FreeFoldM.