Control.Foldl.Monadic.End

Description

This module provides a version of FoldM that can return early

Documentation

newtype FoldEM m a b

FoldM left or return an early result

Constructors

FoldEM

Fields

runFoldEM :: FoldM (ExceptT b m) a b

toCofreeReaderM :: Monad m => FoldEM m a b -> ExceptT b m (Cofree (ReaderT a (ExceptT b m)) b)

Unfold the successive applications of a FoldEM into a Cofree ReaderT ExceptT

toCofreeReaderT :: Monad m => FoldEM m a b -> ExceptT b m (Cofree (ReaderT a (ExceptT b m)) b)

Unfold a FoldEM into a Cofree ReaderT

foldEMToFree :: Monad m => FoldEM m a b -> FreeFoldM (ExceptT b m) a b

Convert a FoldE to a FreeFoldM using toCofreeReaderT

foldEMThenFree :: Monad m => FoldEM m a b -> (b -> FoldEM m a c) -> FreeFoldM (ExceptT c m) a c

using foldFreeThenFreeM instead of directly recursing

foldEMThenEMRec :: Monad m => FoldEM m a b -> (b -> FoldEM m a c) -> FoldEM m a c

Using FreeFoldM, fold with the first FoldEM until it returns, then use its result to generate the second FoldEM and continue.

liftFoldE :: Monad m => FoldE a b -> FoldEM m a b

Lift a FoldE to a FoldEM by returning the results

fromFoldEM :: Monad m => FoldEM m a b -> FoldM m a b

We can convert a FoldEM to a FoldM (without the ExceptT provided by runFoldEM) by including the Either in the state and simply extracting the early return value.

liftFoldEM :: MonadError b m => FoldM m a b -> FoldEM m a b

We can lift a FoldM to a FoldEM without the need to explicitly specify the required ExceptT, by using liftExceptT.

liftExceptT :: MonadError e m => m a -> ExceptT e m a

Lift an instance of MonadError to ExceptT using catchError.

joinExceptT :: Functor m => ExceptT a m a -> m a

Join the error and result terms of an ExceptT