Control.Foldl.Utils

Description

This module provides utilities for working with Fold and FoldM

Documentation

extractM :: Monad m => FoldM m a b -> m b

Extract the monadic result from a FoldM

pureM :: Monad m => m b -> FoldM m a b

Produce a FoldM that ignores all input and returns the given monadic value.

toFunction :: Fold a b -> a -> b

Convert a Fold to a function

toFunctionM :: Monad m => FoldM m a b -> a -> m b

Convert a FoldM to a function

stepFold :: a -> Fold a b -> Fold a b

Perform a single step of a Fold

stepFoldM :: Monad m => a -> FoldM m a b -> FoldM m a b

stepFold, generalized to FoldM

unwrapFoldM :: Applicative m => FoldM m a b -> m (FoldM m a b)

Expose the monadic context of the current state of a FoldM. Can be used, for example, to catch errors before continuing a fold.

innerJoinFoldM :: Monad m => FoldM m a (m b) -> FoldM m a b

Join a monadic result into the monadic context of the FoldM

generalizeMaybe :: Monad m => Fold a (Maybe b) -> FoldM (MaybeT m) a b

Generalize a Fold returning Maybe to MaybeT.

For example:

 λ> F.foldM (generalizeMaybe F.head) [1..10] :: MaybeT Identity Int
 MaybeT (Identity (Just 1))

count :: Enum i => (a -> Bool) -> Fold a i

Count the number of elements that match a predicate

Should be equivalent to: length . prefilter p

countM :: (Enum i, Monad m) => (a -> m Bool) -> FoldM m a i

count using a monadic predicate

Orphan instances

Monad m => Extend (FoldM m a)
duplicated = duplicateM

Instance details

Methods

duplicated :: FoldM m a a0 -> FoldM m a (FoldM m a a0)

extended :: (FoldM m a a0 -> b) -> FoldM m a a0 -> FoldM m a b