Control.Monad.Rep

Description

This module defines a version of Representable where the function tabulated/indexed returns a monadic result.

In other words, instead of providing an isomorphism with Reader, we provide one with ReaderT for a fixed Monad.

Type class

class DistributiveM m f => RepresentableM m f | f -> m where

A Functor f is RepresentableM for a given Monad m if tabulateM and indexM witness an isomorphism to ReaderT (RepM f) m.

 tabulateM . indexM == id
 indexM . tabulateM == id
 tabulateM . return . return == return

Every RepresentableM Functor whose base Monad is a right adjoint is also a right adjoint.

Inuitively, a RepresentableM Functor is a Representable Functor on the Kleisli category of the base Monad.

I'd expect some superclass requirement, but it ends up looking like just Applicative f:

class Applicative f => DistributiveA f where
  distributeM :: Traversable t => t (f a) -> f (t a)

instance Applicative m => DistributiveM (ReaderT a m) where
  distributeM = ReaderT . fmap sequenceA . distribute . fmap runReaderT

Associated Types

type RepM f :: *

If no definition is provided, this will default to GRepM.

Methods

tabulateM :: (RepM f -> m a) -> f a

 fmap (fmap f) . tabulateM == tabulateM . fmap f

If no definition is provided, this will default to gtabulateM.

tabulateM :: (Generic1 f, GRepM f ~ RepM f, GTabulateM m (Rep1 f)) => (RepM f -> m a) -> f a

 fmap (fmap f) . tabulateM == tabulateM . fmap f

If no definition is provided, this will default to gtabulateM.

indexM :: f a -> RepM f -> m a

If no definition is provided, this will default to gindexM.

indexM :: (Generic1 f, GRepM f ~ RepM f, GIndexM m (Rep1 f)) => f a -> RepM f -> m a

If no definition is provided, this will default to gindexM.

Functor

fmapRepM :: RepresentableM m f => (a -> b) -> f a -> f b

fmap inside the Monadic result of a RepresentableM

Distributive

distributeRepM :: (RepresentableM m f, Distributive m, Functor w) => w (f a) -> f (w a)

collect after indexMing

collectRepM :: (RepresentableM m f, Distributive m, Functor w) => (a -> f b) -> w a -> f (w b)

collect after indexMing the result of the provided function

DistributiveM

distributemRepM :: RepresentableM m f => m (f a) -> f (m a)

distributem for RepresentableM

collectmRepM :: RepresentableM m f => (a -> f b) -> m a -> f (m b)

collect for RepresentableM

Apply/Applicative

apRepM :: RepresentableM m f => f (a -> b) -> f a -> f b

Uses the isomorphism

pureRepM :: RepresentableM m f => a -> f a

tabulateM ignoring the index and returning the value

liftRM2 :: RepresentableM m f => (a -> b -> c) -> f a -> f b -> f c

Uses the isomorphism

liftRM3 :: RepresentableM m f => (a -> b -> c -> d) -> f a -> f b -> f c -> f d

Uses the isomorphism

Bind/Monad

bindRepM :: RepresentableM m f => f a -> (a -> f b) -> f b

Equivalent to:

 \m f -> runReaderT $ toReaderT m >>= toReaderT . f

MonadFix

mfixRepM :: (RepresentableM m f, MonadFix m) => (a -> f a) -> f a

Equivalent to:

 runReaderT . mfix . fmap toReaderT

MonadZip

mzipWithRepM :: (RepresentableM m f, MonadZip m) => (a -> b -> c) -> f a -> f b -> f c

indexM, mzipWith, then tabulateM

mzipRepM :: (RepresentableM m f, MonadZip m) => f a -> f b -> f (a, b)

indexM, mzip, then tabulateM

MonadReader

fromReaderT :: RepresentableM m f => ReaderT (RepM f) m a -> f a

RepresentableM m f is isomorphic to ReaderT (RepM f) m a:

 fromReaderT . toReaderT == id
 toReaderT . fromReaderT == id

toReaderT :: RepresentableM m f => f a -> ReaderT (RepM f) m a

The other half of the isomorphism, see: fromReaderT

askRepM :: RepresentableM m f => f (RepM f)

Fetch the value of the environment.

(ask specialized to RepresentableM)

asksRepM

Arguments

:: RepresentableM m f

=> (RepM f -> a)
The selector function to apply to the environment.

-> f a

Retrieve a function of the current environment.

(asks specialized to RepresentableM)

mapRepM :: RepresentableM m f => (m a -> m b) -> f a -> f b

Transform the computation inside a RepresentableM.

(mapReaderT specialized to RepresentableM)

withRepM

Arguments

:: RepresentableM m f

=> (RepM f -> RepM f)
The function to modify the environment.

-> f a
Computation to run in the modified environment.

-> f a

Execute a computation in a modified environment.

(withReaderT specialized to RepresentableM)

localRepM

Arguments

:: RepresentableM m f

=> (RepM f -> RepM f)
The function to modify the environment.

-> f a
Computation to run in the modified environment.

-> f a

Execute a computation in a modified environment.

(local specialized to RepresentableM)

Extend

duplicatedRepM :: (RepresentableM m f, Semigroup (RepM f)) => f a -> f (f a)

duplicateRepMBy (<>)

extendedRepM :: (RepresentableM m f, Semigroup (RepM f)) => (f a -> b) -> f a -> f b

extendRepMBy (<>)

Comonad

duplicateRepM :: (RepresentableM m f, Monoid (RepM f)) => f a -> f (f a)

duplicateRepMBy mappend

extendRepM :: (RepresentableM m f, Monoid (RepM f)) => (f a -> b) -> f a -> f b

extendRepMBy mappend

extractRepM :: (RepresentableM m f, Monoid (RepM f), Comonad m) => f a -> a

extractRepMBy mempty

Comonad, with user-specified monoid

duplicateRepMBy :: RepresentableM m f => (RepM f -> RepM f -> RepM f) -> f a -> f (f a)

indexM then tabulateM twice and use the provided operation to combine the keys

extendRepMBy :: RepresentableM m f => (RepM f -> RepM f -> RepM f) -> (f a -> b) -> f a -> f b

indexM then tabulateM twice and use the provided operation to combine the keys

extractRepMBy :: (RepresentableM m f, Comonad m) => RepM f -> f a -> a

Since we need to return a raw value, m must be a Comonad

WithIndex

imapRepM :: RepresentableM m f => (RepM f -> a -> b) -> f a -> f b

tabulateM to expose the RepM and then fmap

imapMRepM :: RepresentableM m f => (RepM f -> a -> m b) -> f a -> f b

tabulateM to expose the RepM and then (>>=)

ifoldMapRepM :: forall m f a b. (RepresentableM m f, Foldable f, Monoid b) => (RepM f -> a -> b) -> f a -> b

Indexed foldMap

itraverseRepM :: (RepresentableM m f, Traversable f, Applicative g) => (RepM f -> a -> g b) -> f a -> g (f b)

Indexed traverse

MonadFree

wrapRepM :: RepresentableM m f => m (f a) -> f a

Wrap the base Monad of a RepresentableM

Lifting other operations

liftRepM :: RepresentableM m f => m a -> f a

Lift the base Monad of a RepresentableM

liftCallCC

Arguments

:: RepresentableM m f

=> (((a -> m b) -> m a) -> m a)
callCC on the base Monad.

-> ((a -> f b) -> f a)

-> f a

Lift a callCC operation to the base Monad.

liftCatch

Arguments

:: RepresentableM m f

=> (m a -> (RepM f -> m a) -> m a)
catch on the base monad.

-> f a
Computation to attempt.

-> (RepM f -> f a)
Exception handler.

-> f a

Lift a catchError operation to the base monad.

Generics

gtabulateM :: (Generic1 f, GRepM f ~ RepM f, GTabulateM m (Rep1 f)) => (RepM f -> m a) -> f a

A default implementation of tabulateM in terms of GRepM.

gindexM :: (Generic1 f, GRepM f ~ RepM f, GIndexM m (Rep1 f)) => f a -> RepM f -> m a

A default implementation of indexM in terms of GRepM.

type GRepM f = GRepM' (Rep1 f)

A convenient composition of GRepM' and Rep1

type family GRepM' (f :: * -> *) :: * where ...

Generic RepM, with a customer TypeError for unsupported types.

Equations

GRepM' (f :*: g) = Either (GRepM' f) (GRepM' g)

GRepM' Par1 = ()

GRepM' (M1 i c f) = GRepM' f

GRepM' (f :.: g) = (WrappedRepM f, GRepM' g)

GRepM' (Rec1 f) = WrappedRepM f

GRepM' U1 = Void

GRepM' x = TypeError ((Text "The type " :<>: ShowType x) :<>: Text " is not one of (:*:), (:.:), Par1, M1, Rec1, or U1.")

class Monad m => GTabulateM (m :: * -> *) (f :: * -> *) | f -> m where

The typeclass implementing gtabulateM (as gtabulateM') for Generic types

Minimal complete definition

gtabulateM'

Methods

gtabulateM' :: (GRepM' f -> m a) -> f a

class Monad m => GIndexM (m :: * -> *) (f :: * -> *) | f -> m where

The typeclass implementing gindexM (as gindexM') for Generic types

Minimal complete definition

gindexM'

Methods

gindexM' :: f a -> GRepM' f -> m a

newtype WrappedRepM f

WrappedRepM prevents generic RepresentableM instances for recursive types from sending the typechecker into an infinite loop.

See WrappedRep.

Constructors

WrapRepM

Fields

unwrapRepM :: RepM f

todo :: ()

Cofree, CofreeT instances