Control.Monad.Trans.History

Documentation

newtype HistoryT m a

Adds a history of all effects that can be fetched, run, and reset.

Note: Only effects in m are stored in the history so updates to the HistoryStack StateT variable will be ignored during history replays.

See historyExample and historyExample2 for examples.

See MonadHistory for a MonadTransformer type-class a la mtl

Constructors

HistoryT

Fields

unHistoryT :: StateT (HistoryStack m) m a

runHistoryT :: Monad m => HistoryT m a -> m a

Run a HistoryT computation

hoistHistoryT :: (Functor n, Functor m) => (forall x. m x -> n x) -> (forall y. n y -> m y) -> HistoryT m a -> HistoryT n a

Because HistoryT includes the Functor variable in both the argument and result of the StateT Monad, we need natural transformations to and from n to hoist m -> n

class (Monad m, Monad n) => MonadHistory m n | m -> n where

Monads m that support the features provided by HistoryT; namely: getting, lifting and resetting a HistoryStack of effects in n.

MonadHistory instances should satisfy:

 resetHistory >> getHistory = return mempty
 resetHistory >> liftHistory f >> getHistory = resetHistory >> liftHistory f >> return (singletonHistoryStack f)
 getHistory >>= \h0 -> liftHistory f >> getHistory >>= \h1 -> (h0, h1) = getHistory >>= \h0 -> liftHistory f >> return (h0, pushHistory f h0)

Notes:

The second rule can be derived from the first and last.

Only effects in n are stored in the history

Minimal complete definition

getHistory, liftHistory, resetHistory

Methods

getHistory :: m (HistoryStack n)

Get the current HistoryStack

liftHistory :: n a -> m a

Lift the effect Monad to the outer Monad

resetHistory :: m ()

Reset the HistoryStack to mempty

runHistoryStack :: MonadHistory m n => HistoryStack n -> m ()

Run a HistoryStack

breakHistory :: (MonadHistory m n, MonadError e n) => e -> m a

Reset the HistoryStack before throwing the error

historyExample :: IO ()

Example with getHistory, runHistoryStack and resetHistory:

 historyExample :: IO ()
 historyExample = runHistoryT $ do
   let p = liftIO . (print :: Int -> IO ())
       ps = liftIO . putStrLn
   p 1
   p 2
   p 3
   history1 <- getHistory
   ps "repeating history up to now"
   runHistoryStack history1
   ps "(nothing else is repeated)"

   ps ""
   ps "Now we reset:"
   resetHistory

   p 42
   ps "^ it's a number"

   history2 <- getHistory
   ps "repeating history up to now"
   runHistoryStack history2
   ps "(nothing else is repeated)"

Results:

 λ> historyExample
 1
 2
 3
 repeating history up to now
 1
 2
 3
 (nothing else is repeated)

 Now we reset:
 42
 ^ it's a number
 repeating history up to now
 42
 ^ it's a number
 (nothing else is repeated)

historyExample2 :: IO ()

Simple example:

 historyExample2 :: IO ()
 historyExample2 =
   runHistoryT $ do
     liftIO $ putStrLn "repeat me 1"
     liftIO $ putStrLn "repeat me 2"
     liftIO $ putStrLn "repeat me 3"
     history1 <- getHistory
     liftIO $ putStrLn "repeating:"
     stackDepth <- lift $ historyStackDepth history1
     liftIO . putStrLn $ unwords [Repeated, show stackDepth, "layers of IO"]

Results:

 λ> historyExample2
 repeat me 1
 repeat me 2
 repeat me 3
 repeating:
 repeat me 1
 repeat me 2
 repeat me 3
 Repeated 3 layers of IO