Control.Nary

Description

This module defines n-ary functions and operations on them.

Documentation

data Ary (n :: N) (a :: *) (b :: *) where

N-ary functions

λ> :t Ary0 ()
Ary0 () :: Ary 'Z a ()

λ> :t Ary $ x -> Ary0 (succ x)
Ary $ x -> Ary0 (succ x) :: Enum b => Ary ('S 'Z) b b

λ> :t Ary . (Ary .) $ x y -> Ary0 [x, y]
Ary . (Ary .) $ x y -> Ary0 [x, y] :: Ary ('S ('S 'Z)) b [b]

Constructors

Ary0 :: b -> Ary Z a b

Ary :: (a -> Ary n a b) -> Ary (S n) a b

unwrapAry :: Ary (S n) a b -> a -> Ary n a b

Unwrap a single layer from a Ary

arity :: Ary n a b -> Proxy n

Get the n (arity) from a Ary

class KnownN n => AryClass n where

Conversions between representations of an n-ary function

Minimal complete definition

toAryClass, fromAryClass

Associated Types

type NAry (n :: N) a b :: *

An n-ary function

Methods

toAryClass :: NAry n a b -> Ary n a b

Convert a specific arity fuinction to a Ary

fromAryClass :: Ary n a b -> NAry n a b

Convert a Ary to a specific arity fuinction

forallAryClass :: KnownN n => proxy n -> Dict (AryClass n)

All KnownN's are AryClass

toAry :: KnownN n => NAry n a b -> Ary n a b

Convert an n-ary function to Ary:

 toAry ()           :: Ary (ToN 0) Int () :: Ary 'Z Int ()
 toAry (_ -> ())   :: Ary (ToN 1) Int () :: Ary ('S 'Z) Int ()
 toAry (_ _ -> ()) :: Ary (ToN 2) Int () :: Ary ('S ('S 'Z)) Int ()

toAryProxy :: KnownN n => proxy n -> NAry n a b -> Ary n a b

Helper to fix n in toAry

fromAry :: KnownN n => Ary n a b -> NAry n a b

Convert Ary to an n-ary function

class KnownN n => FoldAry n where

Conversions between n-ary functions and Folds

Minimal complete definition

naryFoldClass, naryFoldMClass, naryFromFoldClass, naryFromFoldMClass

Methods

naryFoldClass :: Ary n a b -> Fold a (Maybe b)

naryFoldMClass :: Monad m => Ary n a (m b) -> FoldM (MaybeT m) a b

naryFromFoldClass :: Fold a b -> Ary n a b

naryFromFoldMClass :: Monad m => FoldM m a b -> Ary n a (m b)

forallFoldAry :: KnownN n => proxy n -> Dict (FoldAry n)

All KnownN's are FoldAry

naryFold :: KnownN n => Ary n a b -> Fold a (Maybe b)

Convert Ary to a Fold, returning Nothing if there are fewer than n elements folded.

naryFoldM :: KnownN n => Monad m => Ary n a (m b) -> FoldM (MaybeT m) a b

Convert Ary to a FoldM, returning empty if there are fewer than n elements folded.

naryFromFold :: KnownN n => Fold a b -> Ary n a b

Convert a Fold to an arbitrary-sized Ary function

naryFromFoldProxy :: KnownN n => prxy n -> Fold a b -> Ary n a b

Helper to fix n in naryFromFold

naryFromFoldM :: (KnownN n, Monad m) => FoldM m a b -> Ary n a (m b)

Convert a FoldM to an arbitrary-sized Ary function

naryFromFoldMProxy :: (KnownN n, Monad m) => proxy n -> FoldM m a b -> Ary n a (m b)

Helper to fix n in naryFromFoldM

forallDistributiveAry :: KnownN n => proxy n -> Dict (DistributiveAry n)

All KnownN's are DistributiveAry

class KnownN n => DistributiveAry n where

Polymorphic definitions of distribute and collect over N |

Minimal complete definition

distributeAryClass, collectAryClass

Methods

distributeAryClass :: Functor f => f (Ary n a b) -> Ary n a (f b)

distribute for a particular N

collectAryClass :: Functor f => (a -> Ary n b c) -> f a -> Ary n b (f c)

collect for a particular N

class KnownN n => ApplicativeAry n where

Polymorphic definition of pure and (<*>) over N

Minimal complete definition

pureAryClass, apAryClass

Methods

pureAryClass :: b -> Ary n a b

apAryClass :: Ary n a (b -> c) -> Ary n a b -> Ary n a c

pureAryProxy :: KnownN n => proxy n -> b -> Ary n a b

pureAry with a proxy argument to determine the result size.

class KnownN n => BindAry n where

Polymorphic definition of (>>=) over N

Minimal complete definition

bindAryClass

Methods

bindAryClass :: Ary n a b -> (b -> Ary n a c) -> Ary n a c

class KnownN n => ComonadAry n where

Polymorphic definition of extract and duplicate over N

Minimal complete definition

extractAryClass, duplicateAryClass

Methods

extractAryClass :: Monoid a => Ary n a b -> b

extract for a particular N

duplicateAryClass :: Monoid a => Ary n a b -> Ary n a (Ary n a b)

duplicate for a particular N

class KnownN n => ShiftAry n where

Shift an argument of an Ary function in a level

Minimal complete definition

shiftAryClass

Methods

shiftAryClass :: Ary (S n) a (Ary m a b) -> Ary n a (Ary (S m) a b)

shiftAry :: KnownN n => Ary (S n) a (Ary m a b) -> Ary n a (Ary (S m) a b)

Shift an argument of an Ary function in one level

replicateApply :: Ary n a b -> a -> b

Apply an Ary function to a value n times

replicateMApply :: Monad m => Ary n a b -> m a -> m b

Apply an Ary function to a Monadic value by repeating its action.

joinProxy :: forall (proxy :: kp -> *) (proxy' :: ka -> kp) (a :: ka). proxy (proxy' a) -> Proxy a

Join two, possibly different, proxies into a single Proxy

bindProxy :: forall (proxy :: * -> *) (proxy' :: k -> *) (a :: *) (b :: k). proxy a -> (a -> proxy' b) -> Proxy b

(>>=) generalized to two (possibly different) proxies

proxyResult :: (a -> b) -> Proxy b

Convert a function to a Proxy containing its result type

class (Monad m, Arrow a) => ArrowM m a | a -> m where

Arrows that behave like Kleisli arrows

Minimal complete definition

arrM, joinA

Methods

arrM :: (b -> m c) -> a b c

joinA :: a (m b) b