Data.Distributive.Monadic

Description

This module defines Distributive for a fixed Monad instead of an arbitrary Functor.

Type class

class Monad m => DistributiveM m f | f -> m where

This is the categorical dual of Traversable, for a Monad fixed by our choice of f.

If you define one of distributem, collectm you may define the other using collectmByDistributem, distributemByCollectm, respectively.

All MonadTransformers that result in a Monad are instances of DistributiveM, see collectmTrans.

Methods

distributem :: m (f a) -> f (m a)

The dual of sequence, for a fixed Monad

 distributem = collectm id

distributem :: (Generic1 f, GDistributiveM m (Rep1 f)) => m (f a) -> f (m a)

The dual of sequence, for a fixed Monad

 distributem = collectm id

collectm :: (a -> f b) -> m a -> f (m b)

Map then distributem:

 collectm f = distributem . fmap f

collectm :: (Generic1 f, GDistributiveM m (Rep1 f)) => (a -> f b) -> m a -> f (m b)

Map then distributem:

 collectm f = distributem . fmap f

distributem and collectm in terms of each other

distributemByCollectm :: DistributiveM m f => m (f a) -> f (m a)

distributem in terms of collectm

collectmByDistributem :: DistributiveM m f => (a -> f b) -> m a -> f (m b)

collectm in terms of distributem

Utilities defined using DistributiveM

wrapFoldM :: Monad m => m (FoldM m a b) -> FoldM m a b

Implement wrap using distributem and innerJoinFoldM

ConstT

newtype ConstT (m :: * -> *) a

Ignores the first argument. ConstT is neither a Monad transformer nor a Comonad transformer.

Constructors

ConstT

Fields

runConstT :: a

MonadTrans instances

distributemTrans :: (Monad m, MonadTrans t, Monad (t m)) => m (t m a) -> t m (m a)

distributem for any MonadTrans that's a Monad when its base functor is.

collectmTrans :: (Monad m, MonadTrans t, Monad (t m)) => (a -> t m b) -> m a -> t m (m b)

collectm for any MonadTrans that's a Monad when its base functor is.

We lift the m then (>>=) the function to the result. Finally, we return inside of the transformer.

Generics

gdistributem :: (GDistributiveM m (Rep1 f), Generic1 f) => m (f a) -> f (m a)

A default implementation of distributem for Generic1 types

gcollectm :: (GDistributiveM m (Rep1 f), Generic1 f) => (a -> f b) -> m a -> f (m b)

A default implementation of collectm for Generic1 types

class Monad m => GDistributiveM (m :: * -> *) (f :: * -> *) | f -> m where

The typeclass implementing gdistributem (as gdistributem') for Generic1 types

Minimal complete definition

gcollectm'

Methods

gdistributem' :: m (f a) -> f (m a)

gcollectm' :: (a -> f b) -> m a -> f (m b)