Data.Foldable.Free.Church

Description

This module provides a definition of a free Foldable container in terms of Fold, using existential types in a Church encoding.

Data types

newtype FoldableF a

Church encoding of a free Foldable container

Constructors

FoldableF

Fields

runFoldableF :: forall x y. (x -> a -> x) -> x -> (x -> y) -> y

Construction/Destruction

head :: FoldableF a -> Maybe a

The first element of a FoldableFree, if it exists

tail :: FoldableF a -> FoldableF a

All but the first element of a FoldableFree if it's non-null. Otherwise, returns empty

uncons :: FoldableF a -> Maybe (a, FoldableF a)

Return Just the head and tail or Nothing if null

cons :: a -> FoldableF a -> FoldableF a

Prepend an element to a FoldableFree

Conversion

toFoldableF :: Foldable t => t a -> FoldableF a

Convert any Foldable into a FoldableF.

 toList . toFoldableF == toList

toFoldableF' :: Foldable t => t a -> FoldableF a

Strict toFoldableF

foldableF :: Fold a (FoldableF a)

Fold into a FoldableFree

Distributive/Representable

distributeFold :: Functor f => f (Fold a b) -> Fold a (f b)

Uses tabulateFold and indexFold

indexFold :: Fold a b -> FoldableF a -> b

Index a Fold using a FoldableF

tabulateFold :: (FoldableF a -> b) -> Fold a b

Tabulate a Fold using a FoldableF