Data.Foldable.Monadic.Free

Description

This module provides an implementation of a free FoldableM in terms of FoldM.

Data types

newtype FoldableFreeM m a

A free FoldableM container

Constructors

FoldableFreeM

Fields

runFoldableFreeM :: forall x. FoldM m a x -> m x

Construction/Destruction

head :: Monad m => FoldableFreeM m a -> MaybeT m a

The first element of a FoldableFreeM, if it exists

tail :: Monad m => FoldableFreeM m a -> FoldableFreeM m a

All but the first element of a FoldableFreeM if it's non-null. Otherwise, returns empty

uncons :: Monad m => FoldableFreeM m a -> MaybeT m (a, FoldableFreeM m a)

Return Just the head and tail or Nothing if null

cons :: Monad m => a -> FoldableFreeM m a -> FoldableFreeM m a

Prepend an element to a FoldableFreeM

consM :: Monad m => a -> FoldableFreeM m a -> m (FoldableFreeM m a)

Prepend an element to a FoldableFreeM with the monadic effects exposed

Conversion

foldableFreeM :: Monad m => FoldM m a (FoldableFreeM m a)

FoldM into a FoldableFreeM

toFoldableFreeM :: FoldableM m t => t a -> FoldableFreeM m a

Convert any FoldableM into a FoldableFreeM.

 toListT . toFoldableFreeM == toListT

fromCofree :: Monad m => Cofree m a -> FoldableFreeM m a

Sequence the side effects and then collect the values using consM

toCofreeMaybe :: Monad m => FoldableFreeM m a -> MaybeT m (Cofree (MaybeT m) a)

Convert to a Cofree MaybeT

fromCofreeMaybe :: Monad m => Cofree (MaybeT m) a -> FoldableFreeM m a

Convert a Cofree MaybeT to a FoldableFreeM

cofreeMaybeT :: Monad m => FoldM m a (Maybe (Cofree (MaybeT m) a))

Fold monadically into a Cofree MaybeT

Wrapping/unwrapping

wrapFoldableFreeM :: Monad m => m (FoldableFreeM m a) -> FoldableFreeM m a

Wrap a layer of monadic context into a FoldableFreeM

wrapMaybeT :: Monad m => MaybeT m (FoldableFreeM m a) -> FoldableFreeM m a

Apply wrapFoldableFreeM and replace Nothing with empty

unwrapFoldableFreeM :: Monad m => FoldableFreeM m a -> m (FoldableFreeM m a)

Unwrap a layer of monadic context from a FoldableFreeM

Distributive/Representable

distributeFoldM :: (Monad m, Distributive m, Functor f) => f (FoldM m a b) -> FoldM m a (f b)

Uses tabulateFoldM and indexFoldM. See Distributive.

indexFoldM :: Monad m => FoldM m a b -> FoldableFreeM m a -> m b

Index a FoldM using a FoldableFreeM. See Representable.

tabulateFoldM :: Monad m => (FoldableFreeM m a -> m b) -> FoldM m a b

Tabulate a FoldM by building up successive FoldableFreeMs

MonadFree

wrapFoldM :: Monad m => m (FoldM m a b) -> FoldM m a b

Wrap a layer of the Monadic context into a FoldM. See MonadFree

Orphan instances

Monad m => MonadFree m (FoldM m a)

Instance details

Methods

wrap :: m (FoldM m a a0) -> FoldM m a a0

Monad m => RepresentableM m (FoldM m a)

Instance details

Associated Types

type RepM (FoldM m a) :: *

Methods

tabulateM :: (RepM (FoldM m a) -> m a0) -> FoldM m a a0

indexM :: FoldM m a a0 -> RepM (FoldM m a) -> m a0

Eq1 m => Eq1 (ListT m)

Instance details

Methods

liftEq :: (a -> b -> Bool) -> ListT m a -> ListT m b -> Bool

Monad m => Monad (FoldM m a)

Instance details

Methods

(>>=) :: FoldM m a a0 -> (a0 -> FoldM m a b) -> FoldM m a b

(>>) :: FoldM m a a0 -> FoldM m a b -> FoldM m a b

return :: a0 -> FoldM m a a0

fail :: String -> FoldM m a a0