Data.Foldable.NonEmpty

Description

This module provides a class for folding over non-empty containers, i.e. a subclass of Foldable.

Documentation

class Foldable t => Foldable1 t where

Those data-types that can be folded over and are non-empty

Minimal complete definition

fold1 | foldM1

Methods

fold1 :: Fold1 a b -> t a -> b

Apply a pure strict left fold to a non-empty collection

foldM1 :: Monad m => FoldM1 m a b -> t a -> m b

Apply a Monadic strict left fold to a non-empty collection

head1 :: t a -> a

The first element of a non-empty collection

last1 :: t a -> a

The last element of a non-empty collection

pureFold1Maybe :: Foldable t => Fold1 a b -> t a -> MaybeT m b

Return empty unless the container being folded over is non-empty. Otherwise behaves like fold1.

undefined

impureFold1Maybe :: Foldable t => Fold1 a b -> t a -> MaybeT m b

Return empty unless the container being folded over is non-empty. Otherwise behaves like foldM1.

undefined

asPure :: Monad m => FoldM m a b -> Fold a (m b)

Convert a FoldM to a Fold that returns a Monadic value

asPure1 :: Monad m => FoldM1 m a b -> Fold1 a (m b)

Convert a FoldM1 to a Fold1 that returns a Monadic value

asPureE :: Monad m => FoldEM m a b -> FoldE a (m b)

Convert a FoldEM to a FoldE that returns a Monadic value

undefined

asPureE1 :: Monad m => FoldEM1 m a b -> FoldE1 a (m b)

Convert a FoldEM1 to a FoldE1 that returns a Monadic value

undefined