Data.Functor.NonEmpty

Description

This module provides a Functor transformer that is guaranteed to be non-empty by construction.

Data types

newtype NonEmptyT f a

A Functor with a guranteed element, making it non-empty

Constructors

NonEmptyT

Fields

runNonEmptyT :: (a, f a)

Utilities

extractNonEmptyT :: NonEmptyT f a -> a

Extract the guraranteed value from a NonEmptyT

dropNonEmptyT :: NonEmptyT f a -> f a

Extract all but the guaranteed value from a NonEmptyT

toNonEmpty :: Foldable f => NonEmptyT f a -> NonEmpty a

Convert to NonEmpty

fromNonEmpty :: NonEmpty a -> NonEmptyT [] a

Convert from NonEmpty