Data.History.Stack

Documentation

data HistoryStack f

A stack of effects

Constructors

PushHistory (f (HistoryStack f))

EmptyHistory

hoistHistoryStack :: Functor g => (forall x. f x -> g x) -> HistoryStack f -> HistoryStack g

Hoist one Functor to another within a HistoryStack

wrapHistoryStack :: f (HistoryStack f) -> HistoryStack f

Wrap a layer around a HistoryStack

pushHistory :: Functor f => f a -> HistoryStack f -> HistoryStack f

Push a layer onto a HistoryStack (at the end, i.e. most recent)

singletonHistoryStack :: Functor f => f a -> HistoryStack f

A HistoryStack with a single layer of effects

foreverHistoryStack :: Functor f => f a -> HistoryStack f

Infinite HistoryStack formed by repeating a single layer of effects

cutoffHistoryStack :: Functor f => Int -> HistoryStack f -> HistoryStack f

Cut off a HistoryStack to be at most the input deep

We should have:

 historyStackDepth (cutoffHistoryStack n xs) <= n

flattenHistoryStack :: Monad m => HistoryStack m -> m ()

Flatten a HistoryStack into a single layer

historyStackDepth :: Monad m => HistoryStack m -> m Int

Get the depth of a HistoryStack while evaluating it a la flattenHistoryStack

catchFreeHistoryStack :: MonadError e m => HistoryStack m -> Free m e

At each layer, attempt to catchError and return Pure if one is caught. Otherwise, repeat all of the layers of the HistoryStack indefinitely.

Note: Since we've already caught any possible errors at each layer, no layer of m in the resulting Free m e is throwError.

historyStackToFree :: Functor m => HistoryStack m -> Free m ()

A HistoryStack is isomorphic to the Free Monad of a Functor, specialized to the unit type `()`.

See freeToHistoryStack for the other half of the isomorphism.

freeToHistoryStack :: Functor m => Free m a -> HistoryStack m

The Free Monad of a Functor, specialized to the unit type `()`, is isomorphic to a HistoryStack.

See historyStackToFree for the other half of the isomorphism.