Data.IntPosMap

Documentation

maxDiffSeries :: [Int]

scanIntPosMap :: Enum a => [a] -> [IntPosMap]

Scan, stepping an IntPosMap along a non-empty list

scanIntPosMapL :: Enum a => [a] -> [([a], IntPosMap)]

also return current list

intPosMapThenCycles :: [Int] -> [Int] -> [Int]

The depths of a list as it cycles, after an initial non-cycling section

intPosMapCycles :: [Int] -> [Int]

The depths of a list as it cycles

intPosMapFullCycles :: [Int] -> [Int]

Only output depths of full cycles

distinctDepths :: Int -> Int -> IntMap Int

distinctDepths' :: Int -> Int -> IntMap Int

distinctKeyDepths :: Int -> Int -> Map (IntMap Int) Int

distinct0Depths :: Int -> Int -> IntMap Int

d0d2s_0 :: Int -> [Int]

d0d2s_0_18 :: [Int]

d0d2fs_0 :: Int -> [Int]

prop_d0d2s_0 :: Bool

d0d3s_0 :: Int -> [Int]

d0d3s_0_22 :: [Int]

d0d3fs_0 :: Int -> [Int]

prop_d0d3s_0 :: Bool

d0d3s_1 :: Int -> [Int]

d0d3s_1_21 :: [Int]

d0d3fs_1 :: Int -> [Int]

prop_d0d3s_1 :: Bool

d0d4s_0 :: Int -> [Int]

d0d4s_0_21 :: [Int]

d0d4fs_0 :: Int -> [Int]

prop_d0d4s_0 :: Bool

d0d4s_1 :: Int -> [Int]

d0d4s_1_20 :: [Int]

d0d4fs_1 :: Int -> [Int]

prop_d0d4s_1 :: Bool

d0d4s_2 :: Int -> [Int]

d0d4s_2_30 :: [Int]

d0d4fs_2 :: Int -> [Int]

prop_d0d4s_2 :: Bool

d0d5s_0 :: Int -> [Int]

d0d5s_0_20 :: [Int]

d0d5fs_0 :: Int -> [Int]

prop_d0d5s_0 :: Bool

d0d5s_1 :: Int -> [Int]

d0d5s_0_30 :: [Int]

d0d5s_2 :: Int -> [Int]

d0d5IO :: IO ()

d0d6IO :: IO ()

d0d5s :: [[Int]]

printDistinct0Depths :: Int -> IO ()

distinctDepths2 :: IO ()

All the counts of depthds for 2 distinct elements of lengths [0..]

dd2 :: [[Int]]

Output from distinctDepths2

ofDepth :: Int -> Int -> Int -> Forest Int

All distincts whose resulting IntPosMap is of the given depth

data IntPosMap

PosMap specified to Int so that IntMap may be used.

Constructors

IntPosMap

Fields

getIntPosMap :: IntMap PosCxt

intPosMapLen :: !Int

maxDepthRuns :: Int -> IO ()

tallyMaxDepths :: Int -> [(Int, IntMap Int)]

Tally all max depths and return [(length, tally (maxDepths at level of fseries))]

tallyMaxDepthSums :: Int -> [(Int, [Int])]

baseDepths :: Int -> Int -> [([Int], Int)]

C505837424

stack exec -- basic-exe 25900.09s user 5447.72s system 139% cpu 6:15:27.65 total

assertEqualUptoPerms2 :: Int -> Bool

ee' :: [(Int, [Int])]

toBase is an inverse of fromBase: λ> (n->all ((all [0..n]) . liftM2(==)id.liftM2(.)fromBase(flip toBase))[2..n]) 1000 True (5.28 secs, 1,521,236,536 bytes)

tallyMaxDepthSums 2

ee3' :: [(Int, [Int])]

tallyMaxDepthSums 3

ee :: [[Int]]

ee3 :: [[Int]]

ff :: [[Int]]

ff3 :: [[Int]]

tt :: Int -> Int -> [[Int]]

checkAllsecondDiffsDivPredN :: Int -> Int -> [Bool]

middle1d :: Int -> Int -> Int

middle1ds :: Int -> Int -> (Int, Int)

middle1dKeys :: Int -> Int -> IntMap Int

checkAllTallySumsDivisibleByN :: Int -> [(Int, Bool)]

maxDepthForBools :: Pair [(Int, (Int, IntPosMap))]

ppMaxDepthForBools :: IO ()

Print all of the maximum depths of IntPosMap's with examples

intPosMapDepth :: IntPosMap -> Int

The depth of an IntPosMap, i.e. how many times we recurse in a particular IntPosMap

intPosMapDepths :: IntPosMap -> Maxes Int

intPosMapKeyDepths :: IntPosMap -> IntMap Int

maxDepthForN :: Int -> Tree (Int, [Int])

The max depth tree of an IntPosMap for the given number of elements

ppIntPosMap :: IntPosMap -> String

Show by converting maps to {k1 => x1, k2 => x2, ..}

ppIntPosMap' :: IntPosMap -> String

Show using drawForest after converting to a Forest

ppIntPosMap0 :: IntPosMap -> String

ppIntPosMap for the element at 0

ppIntPosMap1 :: IntPosMap -> String

ppIntPosMap for the element at 1

type Tcxt = (Int, Int, Int, Int, Maybe Int)

Tree context for an IntPosMap.

(val, firstIx, lastIx, len, Maybe diff)

See intPosMapToForest

intPosMapToForest :: IntPosMap -> Forest Tcxt

Convert to a Forest for easier analysis

ppIntMap :: (a -> String) -> IntMap a -> String

Pretty-print an IntMap using an explicit show method

pureIntPosMap :: Int -> IntPosMap

A IntPosMap containing a single element

mapIntPosMap :: (Int -> Int) -> IntPosMap -> IntPosMap

Map over the top layer of keys in a PosMap

mapIntPosMapMonotonic :: (Int -> Int) -> IntPosMap -> IntPosMap

Map over the top layer of keys monotonically in a IntPosMap

fromIntPosMap :: IntPosMap -> NonEmpty Int

See fromPosMap

toIntPosMap :: Foldable t => t Int -> Maybe IntPosMap

Convert any non-empty Foldable to Just a IntPosMap by applying stepIntPosMap to pureIntPosMap and then each successive element, or any empty Foldable to Nothing.

toIntPosMap' :: Foldable t => t Int -> IntPosMap

toIntPosMap, but throw an error instead of returning Nothing

stepIntPosMap :: Int -> IntPosMap -> IntPosMap

Add an element to a IntPosMap by incrementing its length and upserting the element and position.

data PosCxt

The position context of an element, containing its first index (offset), last index (for convenience), length (size), and Either Right a recursive IntPosMap to store series more complex than a single-element cycle or Left to indicate that it's a single-element cycle of that period (difference).

maxIx :: Int using maxIx instead of length allows us to quickly read off the index when adding values.. but if we're updating it anyway, we could use the old length as the new maxIx..

Constructors

PosCxt

Fields

firstIx :: !Int

lastIx :: !Int

posCxtLen :: !Int

ixSeries :: !(Either Int IntPosMap)

Left x for all x's

posCxtDepth :: PosCxt -> Int

The depth of an PosCxt, i.e. how many times we recurse in a particular PosCxt

posCxtToTree :: Int -> PosCxt -> Tree Tcxt

Convert to a Tree for easier analysis

leftToMaybe :: Either a b -> Maybe a

Left to Just

rightToList :: Either a b -> [b]

Right to `[x]`

ppPosCxt :: PosCxt -> String

Pretty-print a PosCxt

purePosCxt :: Int -> PosCxt

A PosCxt for a single repetition of an element at the given index.

stepPosCxt :: Int -> PosCxt -> PosCxt

Just added the case for 0, should refactor as needed lated (e.g. could check that it's a singleton earlier)

Add an element index to a PosCxt.

We increment the size and calculate the next decremented difference: (newIx - lastIx) - 1.

If the isSeries is Nothing (i.e. representing a single-element cycle), then we check whether it's continued (in which case we return Nothing) or we need to make a new IntPosMap to represent the series.

toPosCxt :: Foldable t => t Int -> Maybe PosCxt

Convert any non-empty Foldable to a PosCxt by applying stepPosMap to purePosCxt and then each successive element.

posCxtToBools :: PosCxt -> [Bool]

List, incl. offset, of whether or not the value of a PosCxt is currently at that index.

For example, for the PosCxt of 1 in [0,1,1,0,0,1] posCxtToBools would be:

 [False, True, True, False, False, True]

posCxtIndices :: PosCxt -> [Int]

Convert a PosCxt to a list of indices (positions) of its associated element.

type Gapped = MaybeT

maybeIf :: a -> Bool -> Maybe a

If True then return the value, else Nothing.

posCxtToGapped :: a -> PosCxt -> Gapped [] a

Convert a value and a its PosCxt to a list of of Just the value where it occurs in the list and Nothing (or the list ends) otherwise.

Every Nothing preceeds some Just. If there are no more Just's, the output ends.

runTests :: IO Bool

quickCheckAll