Data.List.Utils

Documentation

diff :: Num a => [a] -> [a]

monoRecordsOn :: Eq b => (a -> b) -> [a] -> [a]

For a monotonically increasing function/series, only output new records.

E.g.

 monoRecordsOn id [1,1,1,2,2,2,2,2,2,3,3,3,3,3,3,3] = [1,2,3]

eventually101 :: [Int] -> (Bool, [Int], Int)

Is the list eventually1s or eventually01s, and True if eventually1s

eventually01s :: [Int] -> ([Int], Int)

The number of elements of the list ending in 0, 1, 0, 1.. and the rest of the list

eventually1s :: [Int] -> ([Int], Int)

The number of elements of the list ending in 1, 1, 1, 1.. and the rest of the list

countPrefixOf :: (a -> Bool) -> [a] -> (Int, [a])

Count the number of elements in the prefix satisfying the predicate and return the rest of the list

testCountPrefixOf :: Int -> Int -> Bool

Test countPrefixOf

countSuffixOf :: (a -> Bool) -> [a] -> ([a], Int)

Count the number of elements in the suffix satisfying the predicate and return the rest of the list

testCountSuffixOf :: Int -> Int -> Bool

Test countSuffixOf

middleOne :: Int -> Int -> [Int]

runs :: Eq a => [a] -> [(Int, a)]

Return the length of the run and the element. For example:

 λ> runs [1,1,2,3,3,3]
 [(2,1),(1,2),(3,3)]

tallyOrd :: (Foldable t, Ord a) => t a -> Map a Int

Tally a Foldable of an Ord

tally :: Foldable t => t Int -> IntMap Int

Tally a Foldable of Ints

tallyMaxes :: Foldable t => t (Maxes Int) -> IntMap Int

Tally Maxes of Ints

toBase :: Integral a => a -> a -> [a]

Convert x to base b

fromBase :: Num a => a -> [a] -> a

Convert a base-b list of digits to a Num

simplePerm :: Integral a => a -> a -> a -> a

Given the number of values, the index of the rotation, and a value to permute (by cycling), perform the permutation.

λ> simplePerm 5 0 $ [1..5] [1,2,3,4,5]

λ> simplePerm 5 1 $ [1..5] [2,3,4,5,1]

λ> simplePerm 5 2 $ [1..5] [3,4,5,1,2]

asserts :: [Bool] -> IO ()

Assert each in list and throw an error with the index if any False

diffUntilDec :: (Num a, Ord a) => [a] -> [a]

diff until the list is anywhere decreasing

binTransform :: Num a => [a] -> [a]

binTransformN :: Num a => [a] -> Int -> [a]

binTransform n times

btrans :: [Int] -> [Int]

Alternate version of binTransform

invBtrans :: [Int] -> [Int]

Alternate version of inverse of binTransform

count :: Foldable t => (a -> Bool) -> t a -> Int

Count the number of elements satisfying the predicate

mapEvens :: (a -> a) -> [a] -> [a]

Map even indices

mapOdds :: (a -> a) -> [a] -> [a]

Map odd indices

parityMap :: (a -> a) -> (a -> a) -> [a] -> [a]

parityMap f g = mapEvens f . mapOdds g = mapOdds g . mapEvens f

evens :: [a] -> [a]

Even-index elements

odds :: [a] -> [a]

Odd-index elements

nest :: (Ord t1, Num t1) => t1 -> (t2 -> t2) -> t2 -> t2

Apply the function the given number of times

sums :: Num a => [a] -> [a]

Sums of successive pairs of elements

data Maxes a

The maximum, along with the number of times that it has occurred

Constructors

Maxes

Fields

numMaxes :: !Int

getMaxes :: a

maxes :: a -> Maxes a

Convert a number to Maxes

unMaxes :: Maxes a -> (Int, a)

Convert Maxes to a tuple

foldl1Maybe :: Foldable t => (b -> a -> b) -> (a -> b) -> t a -> Maybe b

Fold at least one element or fail with Nothing. Use that element to build the first b.

timed :: IO () -> IO Double

Time (in seconds) that an action takes (does not account for timing overhead)

mapM_Timed :: (a -> IO ()) -> [a] -> IO ()

mapM_ with timing info for each action