Data.N

Description

This module provides an implementation of type-level natural numbers and a KnownN that allows us to derive that c :: N -> Constraint holds for all N when we know that c 'Z and forall (n :: N). c n => c ('S n).

See forallN for more details.

Documentation

data N

Type level natural numbers.

Why not use Nat instead? I don't think forallN would be possible.

Constructors

S N

Z

type family Pred (n :: N) :: N where ...

The previous N or Z if Z

Equations

Pred Z = Z

Pred (S n) = n

class KnownN (n :: N) where

Unpack a type-level N into either Z or S n for n ~ Pred n.

This class alone is sufficient to implement many operations, since it allows pattern matching on a type-level natural with resolution of the type to its case when you match on the result.

Minimal complete definition

unpackN, unpackPredN

Methods

unpackN :: proxy n -> Either (n :~: Z) (n :~: S (Pred n))

Unpack a type-level N into either an equivalence with Z or one with S (Pred n). Performed recursively, this allows one to unpack a type-level N completely.

unpackPredN :: proxy n -> Dict (KnownN (Pred n))

Any time we can unpack a N, we can unpack its predecessor.

(It's easy to see this from within the KnownN instances, the implementations are just Dict.)

forallN :: forall (c :: N -> Constraint) (n :: N) (proxy :: N -> *). KnownN n => Dict (c Z) -> (forall (n' :: N). c n' :- c (S n')) -> proxy n -> Dict (c n)

Using forallN, we can convert KnownN to any inductive class constraint on N.

Use like so:

 class ToNatural (n :: N) where
   toNatural_ :: proxy n -> Natural

 instance ToNatural 'Z where
   toNatural_ _ = 0

 instance ToNatural n => ToNatural ('S n) where
   toNatural_ proxyN = 1 + toNatural_ (predProxy proxyN)

 forallToNatural :: KnownN n => proxy n -> Dict (ToNatural n)
 forallToNatural _ = forallN Dict (Sub Dict)

 toNatural :: KnownN n => proxy n -> Natural
 toNatural proxyN = forallToNatural proxyN `withDict` toNatural_ proxyN

In other words, this converts KnownN to any class constraint that's defined inductively over N.

If you have trouble defining the final method since the (n :: N) is only contained in the result, try this method:

 naryFooProxy :: KnownN n => prxy n -> a -> f n a
 naryFooProxy prxy x = forallFoo prxy `withDict` fooClass x

 naryFoo :: KnownN n => a -> f n a
 naryFoo = naryFooProxy Proxy

Conversion to/from Nat

type family ToN (n :: Nat) :: N where ...

Conversion from Nat to N

Equations

ToN 0 = Z

ToN n = S (ToN (n - 1))

toN :: proxy n -> Proxy (ToN n)

Convert a type-level Nat to a type-level N

type family FromN (n :: N) :: Nat where ...

Conversion from N to Nat

Equations

FromN Z = 0

FromN (S n) = 1 + FromN n

fromN :: proxy n -> Proxy (FromN n)

Convert a type-level N to a type-level Nat

Proxy helpers

predProxy :: proxy (S n) -> Proxy n

The precessor of a type-level N

succProxy :: proxy n -> Proxy (S n)

The the successor of a type-level N

unpackPred :: proxy n -> KnownN n :- KnownN (Pred n)

unpackPred works for all n

unpackSucc :: proxy n -> KnownN n :- KnownN (S n)

unpackSucc works for all n