Data.PosMap

Documentation

todo :: ()

Todo:

 TODO:
 - Cycle extraction
 - Cycle-finding: extractCycle . toPosMap :: (Foldable t, Ord a) => t a -> Maybe (Cycle a) -- Cycle a = NonEmpty a
 - Testing: vs naive implementation, testing bijections, etc.
 - Benchmarking: compare to naive, known algorithms. See: https://en.wikipedia.org/wiki/Cycle_detection
 - Add obvious ways of parallelizing the algorithms

note :: ()

Note:

 One of the main attractions of this algorithm is that it's clearly in map-reduce form:
   Convert all partitions of the series to PosMap's or IntPosMap's (set-parallel)
   Combine in parallel (ordered-tree-parallel) (See Semigroup instances)
   Consume in parallel (ordered-tree-parallel) (See conversion to NonEmpty, Seq, Cycle)

 Where:
   set-parallel: Structurally equivalent to mapping over a set with no hidden dependencies
   ordered-tree-parallel: Structurally equivalent to evaluating an ordered tree with no hidden dependencies (e.g. a rose tree: Tree)

 start with a single value, providing a one-layer cycle-decomposition:
   [x]
   { x => (0, 0, 0, Cycled 1) }

 if we add an equal value, we get:
   [x,x]
   { x => (0, 1, 1, Cycled 2) }

 if we add an unequal value, we get:

 t=[x, y]
 t={ x => (0, 0, 0, Cycled 1)
   , y => (1, 1, 0, Cycled 1)
   }

 t=[x, x, y]
 t={ x => (0, 1, 1, Cycled 2)
   , y => (2, 2, 0, Cycled 1)
   }

 t=[x, x, y, y]
 t={ x => (0, 1, 1, Cycled 2)
   , y => (2, 3, 1, Cycled 2)
   }

 t=[x, x, y, y, x]
 t={ x => (0, 1, 2, { 0 => (0, 1, 1, Cycled 2), 2 => (2, 2, 0, Cycled 1) } )
   , y => (2, 3, 1, Cycled 2)
   }

 t=[x, x, y, y, x, x]
 t={ x => (0, 5, 3, { 0 => (0, 3, 3, { 0 => (0, 1, 1, Cycled 2), 1 => (3, 3, 0, Cycled 1)}), 2 => (2, 2, Cycled 1) } )
   , y => (2, 3, 1, Cycled 2)
   }

 t=[x, x, y, y, x, x, y]
 t={ x => (0, 5, 3, { 0 => (0, 3, 3, { 0 => (0, 1, 1, Cycled 2), 1 => (3, 3, 0, Cycled 1)}), 2 => (2, 2, Cycled 1) } )
   , y => (2, 3, 1, { 0 => (2, 3, 1, Cycled 2), 2 => (6, 6, 0, Cycled 1)})
   }

 Map a (first index, last index, size - 1, Cycled size | IntMap (...)

 - We omit the value inside Cycled since it's already in the PosCxt
 - We implement Cycled  | _ as Maybe

 Diffs are actually (x - y - 1), this makes the minimum key of each intermal map 0 instead of 1.

 Cycled i = replicate i 0

data PosMap (a :: *)

To extract a cycle, we get the largest final cycles from each of the subexpressions. We then combine the cycles (that include beginning indices, etc.)

Constructors

PosMap

Fields

getPosMap :: Map a PosCxt

posMapLen :: !Int

purePosMap :: a -> PosMap a

A PosMap containing a single element

mapPosMap :: Ord b => (a -> b) -> PosMap a -> PosMap b

Map over the top layer of keys in a PosMap

mapPosMapMonotonic :: (a -> a) -> PosMap a -> PosMap a

Map over the top layer of keys monotonically in a PosMap

fromPosMap :: PosMap a -> NonEmpty a

The idea is that we convert each value's PosCxt to a list of Bool's, where an element is True iff the element in the original list is equal to the given value.

We convert True to the value, False to Nothing, then we combine the lists by zipping together the Maybe's with (<|>).

Since, for a PosMap to be valid all elements must be accounted for, we know that all of the elements in the resulting list must be Just.

toPosMap :: (Foldable t, Ord a) => t a -> Maybe (PosMap a)

Convert any non-empty Foldable to a PosMap by applying stepPosMap to purePosMap and then each successive element.

stepPosMap :: Ord a => a -> PosMap a -> PosMap a

Add an element to a PosMap by incrementing its length and upserting the element and position.

posMapToIntPosMap :: Enum a => PosMap a -> IntPosMap

Convert a PosMap to an IntPosMap

intPosMapToPosMap :: Enum a => IntPosMap -> PosMap a

Convert an IntPosMap to a PosMap