Data.PseudoPoly

Description

This module provides two pseudo-polymorphic data-types: PseudoPoly and (:>-)

PseudoPoly

data PseudoPoly a b c where

A Pseudo-polymorphic data-type.

If PseudoPoly a b c is inhabited by some terminating value, then the type b is the same as the type c.

To use this equality in practice, pattern-match on the PseudoPoly a b c to get out the constructor; in the body of the pattern-match, the compiler knows that b ~ c.

Alternatively, you may use the eqPseudoPoly helper to extract the prepositional equality: b :~: c.

For example, here we can pattern match on PseudoPoly to recover the fact that b ~ c:

 foo :: (c -> d) -> PseudoPoly a b c -> (b -> d)
 f (PseudoPoly _) -> f

Why is this useful? Take for example Foldable, which requires a polymorphic data type, but doesn't strictly have to. This allows us to declare a Foldable instance for something like ByteString like so:

 import qualified Data.ByteString.Char8 as C8

 instance Foldable (PseudoPoly C8.ByteString Char) where
   foldr f x (PseudoPoly xs) = C8.foldr f x xs

It looks like magic.

Constructors

PseudoPoly :: a -> PseudoPoly a b b

PseudoPoly helpers

fromPseudoPoly :: PseudoPoly a b c -> a

Extract the value from a PseudoPoly. Since we do not depend on the GADT constraints, we can do so lazily.

eqPseudoPoly :: PseudoPoly a b c -> b :~: c

Extract the equality proof from a PseudoPoly1.

swapPseudoPoly :: PseudoPoly a b c -> PseudoPoly a c b

Symmetry of equality

transPseudoPoly :: (c :~: d) -> PseudoPoly a b c -> PseudoPoly a b d

Apply a prepositional type equality to a PseudoPoly

mapPseudoPoly :: (a -> b) -> PseudoPoly a c d -> PseudoPoly b c d

Map over the value in a PseudoPoly

PseudoPoly1

data (a :>- b) c where

PseudoPoly where the second argument is applied to the first

Constructors

PseudoPoly1 :: a b -> (a :>- b) b

PseudoPoly1 helpers

fromPseudoPoly1 :: (a :>- b) c -> a b

Extract the value from a (:>-). Since we do not depend on the GADT constraints, we can do so lazily.

eqPseudoPoly1 :: (a :>- b) c -> b :~: c

Extract the equality proof from a (:>-).

swapPseudoPoly1 :: (a :>- b) c -> (a :>- c) b

Symmetry of equality

transPseudoPoly1 :: (c :~: d) -> (a :>- b) c -> (a :>- b) d

Apply a prepositional type equality to a (:>-)

mapPseudoPoly1 :: (a c -> b c) -> (a :>- c) d -> (b :>- c) d

Map over the value in a (:>-)