Lib

Description

So we have a directed graph (digraph) that we want to represent a category.

All vertices must have at least one self-loop (id)

Composition of arcs is an arc, i.e. e(a,b) && e(b, c) => e(a, c)

0 is no arc, 1 is arc, 2 is anti-arc

forall i. e(i, i)

Final Result: allDGraphs :: Int -> BTree (IntMap Neighborhood, Matrix Int) Binary tree of choices for uncons: either arc or anti arc

Documentation

triNum :: Int -> Int

Triangular numbers

data Arc

A digraph arc

Constructors

Arc

Fields

arcHead :: !Int

arcTail :: !Int

data Neighborhood

The neighborhood of a vertex

Constructors

Neighborhood

Fields

inAdjacent :: IntSet

outAdjacent :: IntSet

(\\\) :: Neighborhood -> Neighborhood -> Neighborhood

\\ set-wise

pureIn :: Int -> Neighborhood

Just the given vertex in the outAdjacent

pureOut :: Int -> Neighborhood

Just the given vertex in the inAdjacent

addIn :: Int -> Neighborhood -> Neighborhood

Add to inAdjacent

addOut :: Int -> Neighborhood -> Neighborhood

Add to outAdjacent

consNeighborhoods :: Arc -> IntMap Neighborhood -> IntMap Neighborhood

Add an Arc to an IntMap of Neighborhoods

type PdState = StateT PdGraph Maybe

Partially determined graph state

runPdState :: Int -> PdState a -> Maybe a

getPdState :: Int -> PdState a -> Maybe PdGraph

data PdGraph

Partially determined dighraphs

Constructors

PdGraph

Fields

numVertices :: !Int

The number of vertices

adjMatrix :: !(Matrix Int)

Diagonal is 1, 0 -> undetermined, 1 -> arc, 2 -> anti-arc

undeterminedArcs :: Set Arc

neighborhoods :: IntMap Neighborhood

The neighborhoods of each vertex

initPdGraph :: Int -> PdGraph

Initial PdGraph with the given number of vertices

unconsUndetermined :: PdState (Maybe Arc)

Get the next undetermined Arc or Nothing

bothChoices :: PdState (PdGraph, PdGraph)

Both possible next determinations

makeChoices :: [Bool] -> PdState ()

Make a list of choices

toBin :: Int -> [Bool]

Convert to list of bits (most significant bit last)

padWith :: Int -> a -> [a] -> [a]

Pad with a value up to the given length

indexedChoices :: Int -> Int -> PdState ()

Indexed choices based on toBin

allChoices :: PdState [PdGraph]

All choices for a PdGraph

allChoicesN :: Int -> [PdGraph]

All choices for a PdGraph with the given number of vertices

determineArc :: Int -> Int -> Int -> PdState ()

Assumes Arc gotten from unconsUndetermined

chooseArc :: Int -> Int -> PdState ()

Choose for an undetermined Arc to be an arc in the final digraph

chooseAnti :: Int -> Int -> PdState ()

Choose for an undetermined Arc to _not_ be an arc in the final digraph

arcCase :: Int -> Int -> PdState a -> PdState a -> PdState a -> PdState a

propagateChoseArc :: Int -> Int -> PdState ()

propagateChoseAnti :: Int -> Int -> PdState ()

forM_IS :: Applicative f => IntSet -> (Int -> f ()) -> f ()

An implementation of forM_ using foldl'

mapM_IS :: Applicative f => (Int -> f ()) -> IntSet -> f ()

An implementation of mapM_ using foldl'

assertIsAsc :: Ord a => [a] -> [a]

Throws an error if not ascending