Lib

Documentation

diff :: Num a => [a] -> [a]

First differences

nest :: Int -> (a -> a) -> a -> a

Nest a function n times on the input

findingCyclesInSeries :: ()

Finding cycles in series efficiently:

 Make a map from values in the series to the initial offset of the value and the sizes of gaps between each value.
   For example:
     1,1,2,2,2,1,3,1,3,2,1,3,2,1,2,3
   =>
     1 => (0, [0,3,1,2,2])
     2 => (2, [0,0,4,2,1])
     3 => (6, [1,2,3])

 First, note that there is a cycle containing x only if the sizes of the gaps between x's cycle (constant, i.e. one element cycles are valid).

 Next, note that if (x) cycles with period (p), and if the next (p-1) value after the initial (x) also cycle, with periods p1, p2, .. p(p0 - 1),
                    then we may be able to combine the cycles into a single one.
                    In particular, if all other periods divide p0, then it's valid.
                    If any period is a multiple of p0, we need to continue checking for that period. e.g. if p0 = 2, p1 = 4, then they're compatible iff p3 is compatible with p0

 Since the map transformation results in strictly smaller cycle problems, we recurse on those problems to solve them in a similar way.
   (with guaranteed termination)

 We can lazily construct the intermediate solutions recursively for all subproblems while traversing each series.

 The problem then becomes of how lazy evaluation affects the time performance, whether some subproblems can be parallelized, and whether there's a better performance order than on-demand.

 - This looks like comonads, especially the duplication and (co?) recursion
 - We probably also want to store how long the current cycle has been going + how many values we've traversed.
   This should allow computing cycles for any offset from the front at the same time.
     (In the stack of gaps, we also store the number of previous equal, cycling values and their period).

 -- valuePositionDiffMap :: (Foldable t, Ord a) => t a -> Map a (Int, [Int])
 -- valuePositionDiffMap =

 valuePositionDiffMap :: Ord a => Fold a (Map a (Int, [Int]))

 valuePositionDiffMapM :: Ord a => FoldM (State (Map a (Int, [Int]))) a ()

data Indexed a

Strict Int-indexed values

Constructors

Indexed

Fields

indexOf :: !Int

indexed :: a

uncurryi :: (Int -> a -> b) -> Indexed a -> b

uncurry for Indexed

curryi :: (Indexed a -> b) -> Int -> a -> b

curry for Indexed

ifoldLen :: Fold (Indexed a) b -> Fold a (Indexed b)

Convert an indexed fold to one that is not indexed, but also returns the length of the input.

ifoldLenM :: Monad m => FoldM m (Indexed a) b -> FoldM m a (Indexed b)

Convert an indexed monadic fold to one that is not indexed, but also returns the length of the input.

stepPositionDiffMap :: Ord a => Int -> a -> Map a (Indexed [Int]) -> Map a (Indexed [Int])

Add an additional position to a position difference map

positionDiffMap :: Ord a => Fold a (Indexed (Map a (Indexed [Int])))

Fold into a position difference map

newtype PosDiffs a

Position differences resulting from a series of a

Constructors

PosDiffs

Fields

getPosDiffs :: Map a (Indexed [Int])

newtype IntPosDiffs

PosDiffs specialized to integers

Constructors

IntPosDiffs

Fields

getIntPosDiffs :: IntMap (Indexed [Int])

partialOrderCycles :: ()

po-cycles (partial-order cycles): the cycles of the comparisions between successive elements.

class CycleIx (f :: *) a where

Non-empty cycles, composed of an index of rotation and a rotated state.

 Operations are provided to:
 - extractbuild from indexstate
 - map over each
 - rotations (succIxF, predIxF, seekIxF)
 - relative indexing

 Example instances:
 - cyclically indexed list, vector, etc.
 - _

Minimal complete definition

toIxF, ixF, ixState, buildIxF, snocIxF

Associated Types

data IxState f a :: *

data IxF f a :: *

For example:

 data IxF f a = IxF
   { ixFindex :: !Int
   , ixFstate :: IxState f a
   }

Methods

toIxF :: Int -> IxState f a -> IxF f a

Half of an isomophism:

 uncurry toIxF :: (Int, IxState f a) -> IxF f a
 liftM2 (,) ixF ixState :: IxF f a -> (Int, IxState f a)

ixF :: IxF f a -> Int

ixState :: IxF f a -> IxState f a

mapIxF :: (Int -> Int) -> IxF f a -> IxF f a

mapIxState :: (IxState f a -> IxState f a) -> IxF f a -> IxF f a

bimapIxF :: (Int -> Int) -> (IxState f a -> IxState f a) -> IxF f a -> IxF f a

succIxF :: IxF f a -> IxF f a

predIxF :: IxF f a -> IxF f a

seekIxF :: Int -> IxF f a -> IxF f a

headIxF :: IxF f a -> a

indexIxF :: Int -> IxF f a -> a

buildIxF :: a -> IxF f a

snocIxF :: a -> IxF f a -> IxF f a

foldableIxF :: Foldable t => t a -> Maybe (IxF f a)

data Seq1 a

Non-empty sequences

data IntSeq1

Non-empty sequences of Ints

data Cycled a

Constructors

Cycled

Fields

runCycled :: Map a ValueState

cycle :: IxF (Seq1 a) a

cycleLen :: Int

cycleReps :: Int

data IntCycled

Constructors

IntCycled

Fields

runIntCycled :: IntMap ValueState

currentCycle :: IntSeq1

currentCycleLen :: Int

currentCycleReps :: Int

data ValueState

Constructors

ValueState

Fields

firstPos :: !Int

lastPos :: !Int

locPositions :: IntCycled

stepCycled :: Ord a => Cycled a -> a -> Cycled a

undefined

data ProdDiff a

This is a first approximation of a division based "finite differences" data type.

It will need a way to handle division by zero better.

  What about saying that the modulo must be in [0..modulus] for non-zero and [1..] for zero?
  In addition, defining: 0 div 0 = 1?
  Then: prevs, 0, positive, .. -> (0, positive) : ..

  If 0/0 = 1
    Then: prevs, 0, 0, .. -> (1, 0) : ..

  If 0/0 = 0
    Then: prevs, 0, 0, .. -> (0, 0) : ..

 0/0 = 0 might be better.. it might be closer to "continuous"

Constructors

ProdDiff

Fields

prodDiffHead :: [a]

prodDiffTail :: [(a, [a])]

succProdDiffMaybe :: Integral a => ProdDiff a -> Maybe (ProdDiff a)

The next ProdDiff or Nothing if the head is empty

divMods1 :: Integral a => a -> [a] -> [(a, a)]

divMods with known first value (so no Maybe and not returned)

divMods :: Integral a => [a] -> Maybe (a, [(a, a)])

divs and mods of previous values, with an initial value, or Nothing if the list is empty.

predProdDiff :: Num a => ProdDiff a -> ProdDiff a

The previous ProdDiff series or itself if the tail is empty

predProdDiffMaybe :: Num a => ProdDiff a -> Maybe (ProdDiff a)

The previous ProdDiff series or Nothing if the tail is empty

predProdDiffWith :: Num a => a -> [a] -> [a] -> [a]

The previous ProdDiff, using the given list

arithmeticProgressionNotes :: ()

Notes on arithmetic progressions

 = zipWith (x mod' ->

 as = [a0, a1, a2, a3, ..]

 divs = [a1 div a0, a2 div a1, a3 div a2, ..]
 mods = [a1 mod a0, a2 mod a1, a3 mod a2, ..]

 as = [a0, a0 * d1 + mod1, % * d2 + mod2, % * d3 + mod3, ..]

 as :: Set (Positive Integer)
 if sum [1 / a | a <- as] == infinity
 then
 forall (n :: Natural).
   exists (k :: N).
     isArithmeticSeries [a !! (k + i) | i <- [0..n]]

 in other words, if
                   sum . fmap (1 /) == const infinity
                 then
                   there are arithmetic progressions of every length.

 For example, all elements are arithmetic progressions of length 1 and all pairs of elements are arithmetic progressions of length 2.
 

arithProgs :: (Num a, Eq a) => [a] -> [[a]]

Get akk arithmetic progressions of all lengths

arithProgsLen :: (Num a, Eq a) => Int -> [a] -> [[a]]

Check for arithmetic progressions of a given length

isArithProg :: (Num a, Eq a) => [a] -> Bool

Is it an arithmetic progression?

predMaybe :: (a -> Bool) -> a -> Maybe a

Just if the predicate returns True

succPairs :: [a] -> [[a]]

succSeries 2

succSeries :: Int -> [a] -> [[a]]

takeExactly n for every tail in tails and drop tails with fewer values

takeExactly :: Int -> [a] -> Maybe [a]

Take exactly the given number of elements or return Nothing

primes :: [Integer]

An infinite list of prime numbers, using a sieve method

data Universe

Constructors

Universe

Fields

maxVal :: !Int

series :: !IntSet

initialUniverse :: Universe

Initial universe, with series = [1]

nextVal :: Universe -> Int

The next value in a Universe

validNextVal :: Universe -> Int -> Bool

Is the next value valid?

A value is invalid if its addition to the series would result in an arithmetic progression of length 3.

addNextVal :: Int -> Universe -> Universe

Add a value to the series, without checking whether it's valid

stepUniverse :: Universe -> Universe

Find and add the next value to the Universe

universes :: [Universe]

All Universes, starting with the initialUniverse

someFunc2 :: IO ()

someFunc :: IO ()

extendedIdea :: ()

Extended idea:

 Extend idea:
 - Tree defined as:
   step prevTree = traverse over complete trees, finding their minimal values
     for each found, [found..] is a valid extension, forming the next layer of the tree.

 e.g. binary tree, minimal, minimal + 1.

 Maybe it's possible to derive the value from the path.

onlyEvery :: Int -> [a] -> [a]

Only keep one out of every n

exampleMapMTimed_ :: IO ()

Example with mapMTimed_

mapMTimed_ :: MonadIO m => (a -> m ()) -> [a] -> m ()

mapM_ where results are timed and the timing is printed with the results

timed :: MonadIO m => m () -> m NominalDiffTime

Run a monadic action and return its duration

rec3 :: Integer -> Integer

Third implementation

 rec0 n = case divMod n 2 of
            (n', 0) ->

 a(2k + 2) = a(2k + 1) + 1, a(2^k + 1) = 2*a(2^k).

 it appears that (rec3 x mod 3^i) cycles with period 2^i.

 i.e.,
   rec3 x mod 3^i == rec3 (mod x 2^i) mod 3^i

notesOnRheRecursion :: ()

Notes on the recursion:

 3 * loop (n div 2) - (2 - mod n 2)
 3 * loop (n div 2) - 2 + mod n 2

 3 * (3 * loop (n div 4) - 2 + mod (n div 2) 2) - 2 + mod n 2
 9 * loop (n div 4) - 6 + 3 * mod (n div 2) - 2 + mod n 2
 9 * loop (n div 4) - 8 + 3 * mod (n div 2) + mod n 2

 9 * (3 * loop (n div 8) - 2 + mod (n div 4)) - (3 * 2 + 2) + 3 * mod (n div 2) + mod n 2
 27 * loop (n div 8) - 9 * 2 + 9 * mod (n div 4)) - (3 * 2 + 2) + 3 * mod (n div 2) + mod n 2
 27 * loop (n div 8) - (9 * 2 + 3 * 2 + 2) + 9 * mod (n div 4) + 3 * mod (n div 2) + mod n 2

 3^i * loop (n div 2^i) - 2 * sum [3^j | j <- [0..i-1]] + sum [3^j * mod (n div 2^j) | j <- [0..i-1]]

 OK
   So it looks like I screwed up the algebra somewhere, but otherwise the theory seems sound.
   It looks like I might be able to travel from minBit to maxBit and multiply the 3's as I go instead of computing powers as well.

 This would result in a log2 time algorithm, with a couple additions, multiplications, bitshifts, etc. per bit.

 (1a + 1b)
 (a + b) / (a * b)
 (a div gcd a b + b div gcd a b) / lcm a b

 suppose:
   n = sum [an j * 2^j | j <- [0..m]]
   where an m = 1
 then
   loop hits 0 at i = m
   loop n = sum [3^j * mod (n div 2^j) | j <- [0..i-1]] - 2 * sum [3^j | j <- [0..i-1]]
   loop n = sum [3^j * (2 - mod (n div 2^j) 2) | j <- [0..i-1]]

   loop n = sum [3^j * mod (n div 2^j) | j <- [0..i-1]] - 2 * ((3^(i - 1) - 1) / 2)
   loop n = sum [3^j * mod (n div 2^j) | j <- [0..i-1]] + 1 - 3^(i - 1)

 a(n)-1 in ternary = n-1 in binary

type MemoT = StateT (IntMap Integer)

Simple implementation of a memoization Monad transformer to memoize a function of the type:

 Int -> Integer

runMemoT :: Monad m => MemoT m a -> m (a, IntMap Integer)

Run MemoT with no initially known values

Return the result along with an IntMap containing all computed values

evalMemoT :: Monad m => MemoT m a -> m a

runMemoT without the IntMap of results

rec3M :: Monad m => Integer -> MemoT m Integer

rec3 memoized with MemoT

arithProgConjecture :: ()

Conjecture on arithmetic progressions

 conjecture:
   if each arithmetic-progression free size has a lower bound of the above form, i.e. f n = k * f (div n (progressionSize - 1)) + line (mod n (progressionSize - 1))
      then since each one forms a lower bound for the minimal series that's ..-free, it forms an uppwer bound for the sum of its reciprocals.
        but since its guaranteed to be at least exponential..

 a(n) = b(n+1) with b(0)=1, b(2n)=3b(n)-2, b(2n+1)=3b(n)-1