Lib

Documentation

foldEitherP :: (a -> Bool) -> Fold a b -> Fold a c -> Fold (Either b c) d -> Fold a d

Assumes the initial is True

data Loc

Source location

Constructors

Loc

Fields

lineNum :: !Int

beginLoc :: !Int

endLoc :: !Int

newtype Token

Constructors

Token

Fields

unToken :: Int

class MonadFail m => MonadToken m where

Make or lookup a Token

Minimal complete definition

token, getToken

Methods

token :: Loc -> Text -> m Token

getToken :: Token -> m (Loc, Text)

data TokenState

Constructors

TokenState

Fields

currentToken :: !Token

lookupTrie :: !(Trie Token)

tokenMap :: !(IntMap (Loc, ByteString))

newtype TokenT m a

Constructors

TokenT

Fields

getTokenT :: StateT TokenState m a

runTokenT :: Monad m => TokenT m a -> m (a, TokenState)

parseText :: Monad m => Text -> Cofree (ReaderT Char m) a -> Either a (m a)

data Pair a

Strict homogenous pairs

Constructors

Pair

Fields

fstPair :: !a

sndPair :: !a

enumCofree :: (Enum i, Functor m) => Cofree m a -> WriterT i (Cofree m) a

mark :: Functor m => WriterT i (Cofree m) a -> WriterT (Pair i) (Cofree m) a

getMark :: WriterT (Pair i) (Cofree m) a -> Pair i

Extract the current marked range

unMarkToken :: (Enum i, MonadToken m) => i -> Text -> WriterT (Pair i) (Cofree m) a -> m (Token, WriterT i (Cofree m) a)

selectToken :: (Enum i, MonadToken m) => i -> Text -> Pair i -> m Token

selectText :: (Enum i, MonadFail m) => Text -> Pair i -> m Text

takeToken :: (Enum i, MonadToken m) => i -> Text -> ReaderT a m (Cofree (ReaderT a m) Bool) -> WriterT (Pair i) (Cofree m) a -> m (Token, WriterT i (Cofree m) a)

Use a Cofree Reader to take all for a token

dropCofreeReader :: Monad m => ReaderT a m (Cofree (ReaderT a m) Bool) -> WriterT (Pair i) (Cofree m) a -> m (WriterT (Pair i) (Cofree m) a)

Drop using a Cofree Reader to encode a (possibly stateful) predicate

unitReader :: Applicative m => (a -> Bool) -> ReaderT a m (Cofree (ReaderT a m) Bool)

unitReaderM :: Applicative m => (a -> m Bool) -> ReaderT a m (Cofree (ReaderT a m) Bool)

digitReader :: Applicative m => ReaderT Char m (Cofree (ReaderT Char m) Bool)

Example usage of unitReader to read digits

fromList :: Alternative f => [a] -> f (Cofree f a)

foldlChar :: Fold Char b -> Fold Text b

foldlChar' :: Fold Char b -> Fold Text b

groupByF :: (a -> a -> Bool) -> Fold a b -> Fold b c -> Fold a c

Fold into groups, then fold those groups.

Roughly equivalent to:

 fold fy . map (fold fx) . groupBy p == groupByF

groupByF' :: (a -> a -> Bool) -> Fold a b -> Fold b c -> Fold a c

groupByF'' :: (a -> a -> Bool) -> Fold a b -> Fold b c -> Fold a c

spanF :: (a -> Bool) -> Fold a b -> (b -> Fold a c) -> Fold a c

Fold with the first Fold until the predicate returns False, then with the second Fold.

For example:

 λ> fold (spanF (sumF (x - liftA2 (,) (pure x) ((+ x) $ sumF))) $ 1..10

sumF :: Num a => Fold a a

nothingT :: Applicative m => MaybeT m a

Fold as a Tree where pure values are Left and subtrees are folded into Right (ignoring unmatched right (endP) values) treeF :: (a -> Bool) -> (a -> Bool) -> Fold (Either a b) b -> Fold a b treeF beginP endP f = simplify $ extract $ treeFM' beginP endP (const (return ())) (generalize f)

pushTop :: Monad m => a -> NonEmpty a -> m (NonEmpty a)

feedTop :: Monad m => a -> NonEmpty (FoldM m (Either a b) c) -> m (NonEmpty (FoldM m (Either a b) c))

popTop :: MonadReader (FoldM m (Either a b) b) m => NonEmpty (FoldM m (Either a b) b) -> m (NonEmpty (FoldM m (Either a b) b))

pushFeedPop :: MonadReader (FoldM m (Either a b) b) m => (a -> Bool) -> (a -> Bool) -> FoldM m (Either a b) b -> a -> NonEmpty (FoldM m (Either a b) b) -> m (NonEmpty (FoldM m (Either a b) b))

pushFeedPopFold :: MonadReader (FoldM m (Either a b) b) m => (a -> Bool) -> (a -> Bool) -> FoldM m (Either a b) b -> FoldM m a (NonEmpty (FoldM m (Either a b) b))

collect :: Enum i => (a -> Bool) -> Fold (Either a i) b -> Fold a b

collectM_ :: (Enum i, Monad m) => (a -> Bool) -> FoldM m (Either a i) b -> FoldM m a b

collectM :: (Enum i, Monad m) => (a -> m Bool) -> FoldM m (Either a i) b -> FoldM m a b

newtype Spaces

Constructors

Spaces

Fields

numSpaces :: Int

type MaybeSpaces a = Either Spaces a

swapEither :: Either a b -> Either b a

collectSpaces :: Fold (MaybeSpaces Char) b -> Fold Char b

collectSpacesM :: Monad m => FoldM m (MaybeSpaces Char) b -> FoldM m Char b

foldSpaces :: Fold Char b -> Fold (MaybeSpaces b) c -> Fold (MaybeSpaces Char) c

foldSpacesM :: Monad m => FoldM m Char b -> FoldM m (MaybeSpaces b) c -> FoldM m (MaybeSpaces Char) c

collectFoldSpaces :: Fold Char a -> Fold (MaybeSpaces a) b -> Fold Char b

collectFoldSpacesM :: Monad m => FoldM m Char a -> FoldM m (MaybeSpaces a) b -> FoldM m Char b

exampleCountChars :: Fold Char Int

stringF :: Fold Char String

newtype FreeReaderT a m b

Constructors

FreeReaderT

Fields

runFreeReaderT :: FreeT (Reader a) m b

class Monad m => MonadReader a m | m -> a where

Minimal complete definition

withRead

Methods

withRead :: (a -> m b) -> m b

cofreeRead :: MonadReader a m => m (Cofree m a)

hoistFreeReaderT :: Monad m => (forall x. m x -> n x) -> FreeReaderT a m b -> FreeReaderT a n b

feedF :: a -> Fold a b -> Fold a b

feedFM :: Monad m => a -> FoldM m a b -> FoldM m a b

joinFoldM :: Monad m => FoldM m a (m b) -> FoldM m a b

returnFoldM :: Monad m => FoldM m (m a) b -> FoldM m a b

bindFoldM :: Monad m => FoldM m a b -> FoldM m (m a) b

mapMFoldM :: Monad m => (b -> m c) -> FoldM m a b -> FoldM m a c

dimapFoldM :: Monad m => (b -> m a) -> (c -> m d) -> FoldM m a c -> FoldM m b d

extractM :: Monad m => FoldM m a b -> m b

swapMaybe :: a -> Maybe b -> Maybe a

endTreeFM :: Monad m => FoldM m a (Cofree (MaybeT (FoldM m a)) b) -> FoldM m a (Maybe b)

stackNonEmpty :: Monad m => (forall x. (x -> NonEmpty x -> m (NonEmpty x)) -> (NonEmpty x -> m (x, Maybe (NonEmpty x))) -> r) -> r

run the resulting fold over a groupBy-like function and collect the results, including the continuation fold runTreeFMSection :: FoldM m a (Cofree (MaybeT (FoldM m a)) b) -> FoldM m b c -> ((FoldM m a b -> FoldM m b c -> r) -> r) -> FoldM m a (Cofree (MaybeT (FoldM m a)) c)

pushNonEmpty :: Monad m => x -> NonEmpty x -> m (NonEmpty x)

popNonEmpty :: Monad m => NonEmpty x -> m (x, Maybe (NonEmpty x))

printFreeT :: FreeT Identity [] String -> IO ()

A small demonstration of the power of treeF toTreeF :: (a -> Bool) -> (a -> Bool) -> Fold a (FreeT Identity [] a) toTreeF px py = treeF px py $ Fold (x y -> case y of Left y' -> x <> [Pure y' ] Right y' -> x <> [Free (Identity y')] ) [] FreeT

printFreeF :: FreeF Identity String (FreeT Identity [] String) -> IO ()

enumF :: Enum i => Fold (i, a) b -> Fold a b

enumF' :: Enum i => Fold (i, a) b -> Fold a b

countF :: Enum i => Fold a i

preFold :: Fold a b -> Fold (b, a) c -> Fold a c

pre-fold, pass the results of the first Fold as a an argument to the second fold (like a scan):

 preFold fx fy == fold . scanl (fmap fx . join (,))

newtype Fold1 a b

Constructors

Fold1

Fields

runFold1 :: ReaderT a (Fold a) b

groupByFM :: Monad m => (a -> a -> m Bool) -> FoldM m a b -> FoldM m b c -> FoldM m a c

Fold into groups using a monadic action, see groupByF

groupByFM' :: Monad m => (a -> a -> m Bool) -> FoldM m a b -> FoldM m b c -> FoldM m a c

snd3 :: (a, b, c) -> b

foldCofree :: Foldable t => Cofree ((->) (b, a)) b -> t a -> b

Fold using the first Fold until False, then pass the result to the second Fold and continue spanF :: (a -> Bool) -> Fold a b -> (b -> Fold a c) -> Fold a c

fold2 :: Foldable t => (b -> a -> b) -> (b -> (a, a) -> b) -> b -> t a -> b

fold2Cofree :: Foldable t => b -> (a -> b) -> (b -> a -> a -> b) -> t a -> b

fold3Cofree :: Foldable t => b -> (a -> b) -> (a -> a -> b) -> (b -> a -> a -> a -> b) -> t a -> b

someFunc :: IO ()