Lib

Documentation

parts :: (Num a, Ord a, Num p) => a -> a -> p

parts k n is the number of partitions of n elements mutually disjoint subsets of at most k elements

prop_partsDecInputs :: Integer -> Integer -> Property

parts decreasing inputs

parts1 :: (Num a, Ord a, Num p) => a -> a -> p

parts with specific handling for more cases

prop_parts1ToParts2 :: Integer -> Integer -> Bool

parts1 to parts2

prop_partsOnLine1 :: Integer -> Property

parts for equal positive inputs equal to 1

 (x -> (1 x) == (parts x x == 1))

parts2 :: (Num a, Ord a, Num p) => a -> a -> p

parts version 2

prop_parts1ToParts2Mod :: Integer -> Integer -> Bool

parts1 to parts2 with a modulus

parts3 :: (Integral a, Ord a) => a -> a -> a

parts version 3

prop_parts2ToParts3 :: Integer -> Integer -> Bool

parts2 to parts3

parts4 :: (Integral a, Ord a) => a -> a -> a

parts version 4

 p 3 n = p 3 (n - 3) + p 2 (n - 1)
   p 3 (n - 3) = p 3 (n - 6) + p 2 (n - 4) = p 3 (n - 6) + (2 * (n - 4) - 1 + (- 1) ^ (4 * n)) div 4 = p 3 (n - 6) + (2 * n - 8) div 4 = p 3 (n - 6) + (n - 4) div 2
   p 2 (n - 1) = p 2 (n - 3) + p 1 (n - 2) = p 2 (n - 3) + 1 = (2 * (n - 3) - 1 + (- 1) ^ (n - 3)) div 4 = (2 * n - 6 - (- 1) ^ n) div 4
 p 3 n = p 3 (n - 6) + (n - 4) div 2 + (2 * n - 6 - (- 1) ^ n) div 4
 p 3 n = p 3 (n - 6) + (2 * n - 8 + 2 * n - 6 - (- 1) ^ n) div 4
 p 3 n = p 3 (n - 6) + (4 * n - 14 - (- 1) ^ n) div 4

prop_parts4PlusTimes :: Integer -> Integer -> Property

Relation for multiples of arguments

 (x y -> (x > y && 0 == (parts4 x (x + y) == parts4 y (2 * y)))

parts5 :: (Integral a, Ord a) => a -> a -> a

parts version 5

parts6 :: Integer -> Integer -> Integer

parts version 5

Alternate implementation of part5

prop_parts4ToParts5 :: Integer -> Integer -> Bool

part4 to part5

cyclicSeriesEx :: Num a => [a]

Example cyclic series

 scanl(+) c (cycle [a1, a2, .., an])

prop_part5ToCyclicSeriesEx :: Bool

part5 to cyclicSeriesEx

prop_ffToCyclicSeriesEx :: Bool

ff to cyclicSeriesEx

chunks :: Int -> [a] -> [[a]]

Break up a list into chunks of size n, or < n if there are < n elements (remaining) in the list

alternate :: Num a => [a] -> [a]

Negate all numbers at odd indices

autoAlternate :: (Num a, Ord a) => [a] -> [a]

alternate if signums of the first two elements differ

ff :: Integral a => a -> a

See prop_ffToCyclicSeriesEx

toPoly :: (Eq a, Fractional a) => [a] -> Maybe (Poly a)

Convert a list to a Poly if possible

toPolys :: (Eq a, Fractional a) => [a] -> [Poly a]

Convert a list of a list of Polys

toPolyN :: (Eq a, Fractional a) => Int -> [a] -> Maybe (Poly a)

Convert to Poly if possible using the given number of elements

takeExactly :: Int -> [a] -> Maybe [a]

Take exactly the given number of elements or return Nothing

interpolate :: Fractional a => [a] -> Poly a

Interpolate a list of Fractional values into a Poly

data PolyF a where

Polynomials with variables and division

Constructors

Var :: PolyF a

(:+) :: a -> a -> PolyF a

(:-) :: a -> a -> PolyF a

(:*) :: a -> a -> PolyF a

(:/) :: a -> a -> PolyF a

var :: Poly a

Poly variable

newtype Poly a

Univariate polynomials

Constructors

Poly

Fields

unPoly :: Free PolyF a

simplify :: Fractional a => Poly a -> Poly a

Simplify combinations of constants and x / x

wrap2 :: MonadFree m f => (a -> b -> m (f c)) -> a -> b -> f c

wrap after applying a binary function

runPoly :: Fractional a => Poly a -> a -> a

Evaluate a univariate polynomial on the given input

runPolyF :: Fractional a => PolyF a -> a -> a

Run a single Functor layer of a Poly

mapHead1 :: (a -> a) -> NonEmpty a -> NonEmpty a

Map over the head of a NonEmpty list

diff :: Num a => [a] -> [a]

First differences

findCycle :: Eq a => [a] -> Maybe [a]

Find a subsequence ys of xs such that:

 and $ zipWith (==) xs (cycle ys)

findPolyCycle :: Integral a => a -> Maybe ([a], Int)

Find a "poly cycle", i.e. a cycle that results from taking alternate's and diff's of the original series.

Such a cycle can always be reconstituted into a piecewise polynomial, with cases for each residue modulo the period of the cycle.

findPolyCycleN :: Integral a => a -> Int -> Maybe ([a], Int)

findPolyCycle with a given period

exampleFindCycle :: (Show a, Integral a) => a -> Int -> IO ()

findCycle and print, from autoAlternateing parts5

autoDiff :: (Ord a, Num a) => [a] -> [a]

Automatically alternate the first differences

  autoAlternate . diff

nest :: Int -> (a -> a) -> a -> a

Nest a function n times on the input

isCycle :: Eq a => [a] -> [a] -> Bool

Is the second argument equal to:

 take n $ cycle xs

for some n?

boolMaybe :: (a -> Bool) -> a -> Maybe a

Just if the predicate returns True

log2 :: Integral a => a -> a

Integral log base 2

mapMTimed_ :: (a -> IO ()) -> [a] -> IO ()

mapM_ with timing information printed

timed :: IO () -> IO NominalDiffTime

Run and get duration using getCurrentTime

runMain :: IO ()

findPolyCycleN 6 $ [1..]

runTests :: IO Bool