Lib

Documentation

newtype Conj a

A conjunction is simply modeled as a list

Constructors

Conj

Fields

runConj :: [a]

evalConj :: Conj Bool -> Bool

data Upto3 a

Contains [1..3] values

Constructors

Upto1 a

Upto2 a a

Upto3 a a a

type Disj3 a = Upto3 a

Disjunction of up to three values

evalDisj3 :: Disj3 Bool -> Bool

newtype Var a

Variables

 False -> a
 True -> not a

Constructors

Var

Fields

runVar :: (Bool, a)

evalVar :: Var Bool -> Bool

type Sat3 a = Conj (Disj3 (Var a))

3-Sat expressions

evalSat3 :: Sat3 Bool -> Bool

data Universe a

Constructors

Universe

Fields

vars :: Set a

completeVars :: Set a

expr :: Sat3 a

superExprs :: [Sat3 a]

evalUniverse :: Universe Bool -> Bool

pickCompleteVar :: Universe a -> Maybe (a, Universe a)

Pick a var if one exists pickVar :: Universe a -> Maybe (a, Universe a)

Pick a complete var if one exists, removing it from the Set of completeVars

adjacentTo :: Ord a => a -> Sat3 a -> (Set a, Set a)

Partition into variables adjacent and non-adjacent to the given one

adjacentToDisj3 :: Ord a => a -> Disj3 a -> (Set a, Set a)

adjacentTo for Disj3

filterAdj :: Ord a => a -> Sat3 a -> (Sat3 a, Sat3 a)

Partition Sat3 into adjacent and non-adjacent subexpressions

filterByVar :: Ord a => a -> Universe a -> Universe a

undefined

Choose a variable to split the Universe on

adjacentVarsPass :: ()

A pass on adjacent variables in a 3-SAT expression:

 Any vars that are not adjacent to var are removed from vars

 If var is adjacent to any var' that has an instance that is not adjacent to any instance of var,
   then var' is removed from completeVars

 The new expression has all subexpressions that contain var

 The new superExprs' head is all subexpressions that do not contain var

mergeUp :: Ord a => Universe a -> Maybe (Universe a)

undefined

 f universe = case pickCompleteVar universe of
   Nothing -> case mergeUp universe of
     Nothing -> universe
     Just universe' -> f universe'
   Just (var, universe') -> f (filterByVar var universe')